refactor(agent): split agent_linux.go into focused linux files
Move state/stats/buffers, run utilities, BPF wire decoders, BPF attach
starters, ring readers, map integrity watcher, policy/enforce map helpers,
and digest helpers into agent_linux_*.go. Keep Run/Main/loadAllowedDomainsMap
and package doc on agent_linux.go. No behavior change.
