+++ internal/agent/agent_linux.go
@@ -8,20 +8,11 @@
 package agent
 
 import (
-	"bytes"
 	"context"
-	"encoding/binary"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"log/slog"
-	"net"
-	"os"
 	"os/signal"
-	"path/filepath"
-	"runtime/debug"
-	"slices"
-	"strings"
 	"sync"
 	"syscall"
 	"time"
@@ -30,8 +21,6 @@ import (
 	"github.com/cilium/ebpf/features"
 	"github.com/cilium/ebpf/link"
 	"github.com/cilium/ebpf/ringbuf"
-	"github.com/cilium/ebpf/rlimit"
-	"github.com/coldstep-io/coldstep/internal/atomicwrite"
 	"github.com/coldstep-io/coldstep/internal/bpf/tracebpfaudit"
 	"github.com/coldstep-io/coldstep/internal/bpf/traceconnect"
 	"github.com/coldstep-io/coldstep/internal/bpf/tracedns"
@@ -45,2901 +34,7 @@ import (
 	"github.com/coldstep-io/coldstep/internal/proctree"
 	"github.com/coldstep-io/coldstep/internal/report"
 	"github.com/coldstep-io/coldstep/internal/telemetry"
-	"golang.org/x/sys/unix"
 )
-
-type execEvent struct {
-	TGID    uint32
-	TID     uint32
-	Comm    [16]byte
-	ExePath [256]byte
-}
-
-// removeMemlockRlimit is replaceable in tests.
-var removeMemlockRlimit = rlimit.RemoveMemlock
-
-func initMemlock() error {
-	if err := removeMemlockRlimit(); err != nil {
-		return fmt.Errorf("init memlock rlimit: %w", err)
-	}
-	return nil
-}
-
-// bpf_audit events are decoded directly via decodeBPFAuditEvent on the raw
-// ringbuf byte slice; no Go-side struct mirror is required.
-
-type runStats struct {
-	mu                              sync.Mutex
-	execN                           int
-	tcpN                            int
-	udpN                            int
-	httpN                           int
-	tlsN                            int
-	procForkN                       int
-	fsN                             int
-	connect4TupleUpdateFailuresN    int
-	udpRingbufReserveFailuresN      int
-	dnsRingbufReserveFailuresN      int
-	connectRingbufReserveFailuresN  int
-	httpRingbufReserveFailuresN     int
-	tlsRingbufReserveFailuresN      int
-	execRingbufReserveFailuresN     int
-	forkRingbufReserveFailuresN     int
-	fsRingbufReserveFailuresN       int
-	udpSendmsgMultiIovecObservedN   int
-	tlsWritevMultiIovecObservedN    int
-	unobservedEgressSyscallsN       int
-	ioUringSetupObservedN           int
-	tcpDNSResponsesObservedN        int
-	tcpDNSSkippedShortReadN         int
-	bpfAuditN                       int
-	bpfMapIntegrityFailuresN        int
-	bpfDNSCacheUpdateFailuresN      int
-	bpfAuditRingbufReserveFailuresN int
-	bpfHeartbeatFailures            int
-	policyCounts                    map[string]int
-	droppedCounts                   map[string]int
-}
-
-type forkSectionState struct {
-	mu         sync.Mutex
-	readErrors int
-}
-
-func newForkSectionState() *forkSectionState {
-	return &forkSectionState{}
-}
-
-func (s *forkSectionState) addReadError() {
-	s.mu.Lock()
-	s.readErrors++
-	s.mu.Unlock()
-}
-
-type forkSectionSnapshot struct {
-	readErrors int
-}
-
-func (s *forkSectionState) snapshot() forkSectionSnapshot {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return forkSectionSnapshot{readErrors: s.readErrors}
-}
-
-type fsSectionState struct {
-	mu         sync.Mutex
-	readErrors int
-}
-
-func newFSSectionState() *fsSectionState { return &fsSectionState{} }
-
-func (s *fsSectionState) addReadError() {
-	s.mu.Lock()
-	s.readErrors++
-	s.mu.Unlock()
-}
-
-type fsSectionSnapshot struct {
-	readErrors int
-}
-
-func (s *fsSectionState) snapshot() fsSectionSnapshot {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return fsSectionSnapshot{readErrors: s.readErrors}
-}
-
-type fsRowBuffer struct {
-	mu   sync.Mutex
-	max  int
-	rows []report.FSDigestRow
-}
-
-func newFSRowBuffer(max int) *fsRowBuffer { return &fsRowBuffer{max: max} }
-
-func (b *fsRowBuffer) add(r report.FSDigestRow) {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	b.rows = append(b.rows, r)
-	_ = trimRing(&b.rows, b.max) // overflow already accounted for via fs_cap counter in readFSRing.
-}
-
-func (b *fsRowBuffer) snapshot() []report.FSDigestRow {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	cp := make([]report.FSDigestRow, len(b.rows))
-	copy(cp, b.rows)
-	return cp
-}
-
-type forkEdgeBuffer struct {
-	mu        sync.Mutex
-	max       int
-	totalAdds int
-	edges     []proctree.Edge
-}
-
-func newForkEdgeBuffer(max int) *forkEdgeBuffer {
-	return &forkEdgeBuffer{max: max}
-}
-
-func (b *forkEdgeBuffer) add(e proctree.Edge) {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	b.totalAdds++
-	b.edges = append(b.edges, e)
-	_ = trimRing(&b.edges, b.max) // overflow surfaced via forkEdgeBuffer.snapshot() truncation flag.
-}
-
-func (b *forkEdgeBuffer) snapshot() ([]proctree.Edge, bool) {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	return slices.Clone(b.edges), b.max > 0 && b.totalAdds > b.max
-}
-
-type networkSectionState struct {
-	mu sync.Mutex
-
-	tcpReadErrors    int
-	tcpDecodeErrors  int
-	udpReadErrors    int
-	udpDecodeErrors  int
-	httpReadErrors   int
-	httpDecodeErrors int
-	tlsReadErrors    int
-	tlsDecodeErrors  int
-}
-
-type networkSectionSnapshot struct {
-	tcpReadErrors    int
-	tcpDecodeErrors  int
-	udpReadErrors    int
-	udpDecodeErrors  int
-	httpReadErrors   int
-	httpDecodeErrors int
-	tlsReadErrors    int
-	tlsDecodeErrors  int
-}
-
-const (
-	denyProtoTCP                = 1
-	denyProtoUDP                = 2
-	denyReasonDstNotAllowlisted = 1
-	linuxAFInet                 = 2
-
-	// BPF↔Go wire-format size contract. Each constant is paired with a
-	// `_Static_assert(sizeof(struct X) == N)` in the matching bpf/*.c file
-	// so that any drift on either side fails compilation immediately.
-	// Values were determined empirically (clang -target bpf, sizeof()).
-	connectEventWireSize   = 32  // 4+4+16+4+2 fields, aligned to 4 → 32
-	udpSendEventWireSize   = 36  // 4+4+16+4+2+_pad[2]+4 datagram_len → 36
-	httpSniffEventWireSize = 228 // 4+4+16+4+2+_pad[2]+2+payload[192] → 228
-	tlsSniffEventWireSize  = 292 // 4+4+16+4+2+_pad[2]+2+payload[256] → 292
-	execEventWireSize      = 280 // 4+4+16+exe_path[256] → 280
-	forkEventWireSize      = 48  // 4+4+parent_comm[16]+child_comm[16]+4(sid)+4(pidns) → 48
-	fsEventWireSize        = 284 // 4+4+16+1+path[256]+_pad[3] → 284
-	denyEventWireSize      = 46  // packed: 4+4+16+1+1+1+_pad+daddr[16]+dport[2] → 46
-	bpfAuditEventWireSize  = 28  // 4(tgid)+4(tid)+4(cmd)+comm[16] → 28
-	// trace_dns.bpf.c dns_sniff_event: __u32 len + __u8 is_tcp + __u8 _pad[3] + data[DNS_SNIFF_MAX]
-	dnsSniffMaxPayload          = 4096                   // DNS_SNIFF_MAX in trace_dns.bpf.c
-	dnsSniffEventWireSizeLegacy = 4 + dnsSniffMaxPayload // pre-is_tcp layout (__u32 len + data[])
-	dnsSniffEventWireSize       = 4 + 1 + 3 + dnsSniffMaxPayload
-
-	// Header-only sub-sizes used by the http/tls capture decoders to slice
-	// out the payload window. Pair these with the respective WireSize above.
-	httpSniffEventHeaderSize = 34 // 4+4+16+4+2+_pad[2]+2 capture_len
-	tlsSniffEventHeaderSize  = 34 // same layout
-
-	// After the first enforce deny, read additional deny ringbuf records briefly so JSONL/digest
-	// capture a burst (e.g. TCP + UDP) before fail-fast shutdown.
-	enforceDenyDrainMaxEvents = 32
-	enforceDenyDrainDuration  = 1200 * time.Millisecond
-	enforceDenyDrainReadSlice = 50 * time.Millisecond
-
-	// Canary constants matching struct canary_event in trace_connect.bpf.c.
-	canaryMagic         uint32 = 0xCA1A1210
-	canaryEventWireSize        = 16 // 4 magic + 4 pad + 8 seq_nr
-	canaryInterval             = 10 * time.Second
-	canaryTimeout              = 30 * time.Second
-
-	ringReadRetryBaseDelay = 5 * time.Millisecond
-	ringReadRetryMaxDelay  = 200 * time.Millisecond
-)
-
-type ringReadRetryBackoff struct {
-	current time.Duration
-	sleepFn func(time.Duration)
-}
-
-func newRingReadRetryBackoff() *ringReadRetryBackoff {
-	return &ringReadRetryBackoff{sleepFn: time.Sleep}
-}
-
-func (b *ringReadRetryBackoff) nextDelay() time.Duration {
-	if b.current <= 0 {
-		b.current = ringReadRetryBaseDelay
-		return b.current
-	}
-	next := b.current * 2
-	if next > ringReadRetryMaxDelay {
-		next = ringReadRetryMaxDelay
-	}
-	b.current = next
-	return b.current
-}
-
-func (b *ringReadRetryBackoff) sleep() time.Duration {
-	delay := b.nextDelay()
-	if delay <= 0 {
-		return 0
-	}
-	b.sleepFn(delay)
-	return delay
-}
-
-func (b *ringReadRetryBackoff) reset() {
-	b.current = 0
-}
-
-// canaryState tracks telemetry integrity canaries across the BPF
-// ringbuf pipeline. Userspace arms a sequence via the canary_trigger
-// BPF map; the BPF program emits a canary_event into connect_events;
-// the ringbuf reader calls noteReceived. If a canary doesn't arrive
-// within canaryTimeout, the pipeline is considered compromised.
-type canaryState struct {
-	mu           sync.Mutex
-	lastSent     uint64
-	lastSentAt   time.Time
-	lastReceived uint64
-	lastRecvAt   time.Time
-	failCount    int
-}
-
-func newCanaryState() *canaryState { return &canaryState{} }
-
-func (c *canaryState) noteSent(seq uint64) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.lastSent = seq
-	c.lastSentAt = time.Now()
-}
-
-func (c *canaryState) noteReceived(seq uint64) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.lastReceived = seq
-	c.lastRecvAt = time.Now()
-}
-
-type canarySnapshot struct {
-	lastSent       uint64
-	lastReceived   uint64
-	failCount      int
-	pipelineOK     bool
-	lastSentAt     time.Time
-	lastReceivedAt time.Time
-}
-
-func (c *canaryState) snapshot() canarySnapshot {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	ok := true
-	if c.lastSent > 0 && c.lastReceived < c.lastSent &&
-		time.Since(c.lastSentAt) > canaryTimeout {
-		ok = false
-	}
-	return canarySnapshot{
-		lastSent:       c.lastSent,
-		lastReceived:   c.lastReceived,
-		failCount:      c.failCount,
-		pipelineOK:     ok,
-		lastSentAt:     c.lastSentAt,
-		lastReceivedAt: c.lastRecvAt,
-	}
-}
-
-func (c *canaryState) checkAndRecordFailure() bool {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if c.lastSent > 0 && c.lastReceived < c.lastSent &&
-		time.Since(c.lastSentAt) > canaryTimeout {
-		c.failCount++
-		return true
-	}
-	return false
-}
-
-type enforcementState struct {
-	mu                     sync.Mutex
-	mode                   string
-	allowlistSize          int
-	denyCountN             int
-	denyReserveFailuresN   int
-	mapIntegrityFailures   int
-	expectedEntries        int
-	expectedIgnoredEntries int
-	firstDenyRowV          *report.DenyDigestRow
-}
-
-type enforcementSnapshot struct {
-	mode                 string
-	allowlistSize        int
-	denyCount            int
-	denyReserveFailures  int
-	mapIntegrityFailures int
-	firstDeny            *report.DenyDigestRow
-}
-
-type enforceBackendConfig struct {
-	modeEnforce bool
-	haveLSM     bool
-}
-
-type enforceBackendOutcome struct {
-	backend string
-}
-
-const (
-	enforceBackendDetect = "detect"
-	enforceBackendLSM    = "lsm"
-	enforceBackendCgroup = "cgroup"
-
-	enforceModeLSM    = "enforce+lsm"
-	enforceModeCgroup = "enforce+cgroup"
-)
-
-func chooseEnforceBackend(cfg enforceBackendConfig, lsmAttachErr error) enforceBackendOutcome {
-	if !cfg.modeEnforce {
-		return enforceBackendOutcome{backend: enforceBackendDetect}
-	}
-	if cfg.haveLSM && lsmAttachErr == nil {
-		return enforceBackendOutcome{backend: enforceBackendLSM}
-	}
-	return enforceBackendOutcome{backend: enforceBackendCgroup}
-}
-
-func enforceModeForBackend(backend string) string {
-	if backend == enforceBackendLSM {
-		return enforceModeLSM
-	}
-	return enforceModeCgroup
-}
-
-type enforceDenyError struct {
-	protocol string
-	dst      string
-	dport    uint16
-	reason   string
-}
-
-func (e enforceDenyError) Error() string {
-	return fmt.Sprintf("enforce deny: protocol=%s dst=%s dport=%d reason=%s", e.protocol, e.dst, e.dport, e.reason)
-}
-
-func newEnforceDenyError(ev telemetry.DenyEvent) error {
-	return enforceDenyError{
-		protocol: ev.Protocol,
-		dst:      ev.Dst,
-		dport:    ev.Dport,
-		reason:   ev.Reason,
-	}
-}
-
-func isEnforceDenyError(err error) bool {
-	var e enforceDenyError
-	return errors.As(err, &e)
-}
-
-func newEnforcementState() *enforcementState {
-	return &enforcementState{}
-}
-
-func (s *enforcementState) setModeAndAllowlist(mode string, allowlistSize, ignoredSize int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.mode = mode
-	s.allowlistSize = allowlistSize
-	s.expectedEntries = allowlistSize
-	s.expectedIgnoredEntries = ignoredSize
-}
-
-func (s *enforcementState) addMapIntegrityFailure() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.mapIntegrityFailures++
-}
-
-func (s *enforcementState) mapIntegrityFailureCount() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.mapIntegrityFailures
-}
-
-func (s *enforcementState) noteDeny(row report.DenyDigestRow) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.denyCountN++
-	if s.firstDenyRowV == nil {
-		cp := row
-		s.firstDenyRowV = &cp
-	}
-}
-
-func (s *enforcementState) denyCount() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.denyCountN
-}
-
-func (s *enforcementState) firstDeny() *report.DenyDigestRow {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.firstDenyRowV == nil {
-		return nil
-	}
-	cp := *s.firstDenyRowV
-	return &cp
-}
-
-func (s *enforcementState) snapshot() enforcementSnapshot {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	out := enforcementSnapshot{
-		mode:                 s.mode,
-		allowlistSize:        s.allowlistSize,
-		denyCount:            s.denyCountN,
-		denyReserveFailures:  s.denyReserveFailuresN,
-		mapIntegrityFailures: s.mapIntegrityFailures,
-	}
-	if s.firstDenyRowV != nil {
-		cp := *s.firstDenyRowV
-		out.firstDeny = &cp
-	}
-	return out
-}
-
-func (s *enforcementState) setDenyReserveFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.denyReserveFailuresN = n
-}
-
-func newNetworkSectionState() *networkSectionState {
-	return &networkSectionState{}
-}
-
-func (s *networkSectionState) addTCPReaderError() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.tcpReadErrors++
-}
-
-func (s *networkSectionState) addTCPDecodeError() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.tcpDecodeErrors++
-}
-
-func (s *networkSectionState) addUDPReaderError() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.udpReadErrors++
-}
-
-func (s *networkSectionState) addUDPDecodeError() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.udpDecodeErrors++
-}
-
-func (s *networkSectionState) addHTTPReaderError() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.httpReadErrors++
-}
-
-func (s *networkSectionState) addHTTPDecodeError() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.httpDecodeErrors++
-}
-
-func (s *networkSectionState) addTLSReaderError() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.tlsReadErrors++
-}
-
-func (s *networkSectionState) addTLSDecodeError() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.tlsDecodeErrors++
-}
-
-func (s *networkSectionState) snapshot() networkSectionSnapshot {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return networkSectionSnapshot{
-		tcpReadErrors:    s.tcpReadErrors,
-		tcpDecodeErrors:  s.tcpDecodeErrors,
-		udpReadErrors:    s.udpReadErrors,
-		udpDecodeErrors:  s.udpDecodeErrors,
-		httpReadErrors:   s.httpReadErrors,
-		httpDecodeErrors: s.httpDecodeErrors,
-		tlsReadErrors:    s.tlsReadErrors,
-		tlsDecodeErrors:  s.tlsDecodeErrors,
-	}
-}
-
-func newRunStats() *runStats {
-	return &runStats{
-		policyCounts:  make(map[string]int),
-		droppedCounts: make(map[string]int),
-	}
-}
-
-func (s *runStats) addDropped(kind string) {
-	if strings.TrimSpace(kind) == "" {
-		return
-	}
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.droppedCounts[kind]++
-}
-
-func (s *runStats) snapshotDroppedCounts() map[string]int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	out := make(map[string]int, len(s.droppedCounts))
-	for k, v := range s.droppedCounts {
-		out[k] = v
-	}
-	return out
-}
-
-func (s *runStats) addExec() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.execN++
-}
-
-func (s *runStats) addProcFork() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.procForkN++
-}
-
-func (s *runStats) procForkTotal() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.procForkN
-}
-
-func (s *runStats) addFS() {
-	s.mu.Lock()
-	s.fsN++
-	s.mu.Unlock()
-}
-
-func (s *runStats) addPolicyLocked(cl policy.Class) {
-	s.policyCounts[string(cl)]++
-}
-
-func (s *runStats) addTCP(cl policy.Class) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.tcpN++
-	s.addPolicyLocked(cl)
-}
-
-func (s *runStats) addUDP(cl policy.Class) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.udpN++
-	s.addPolicyLocked(cl)
-}
-
-func (s *runStats) addHTTP(cl policy.Class) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.httpN++
-	s.addPolicyLocked(cl)
-}
-
-func (s *runStats) addTLS(cl policy.Class) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.tlsN++
-	s.addPolicyLocked(cl)
-}
-
-func (s *runStats) setConnect4TupleUpdateFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.connect4TupleUpdateFailuresN = n
-}
-
-func (s *runStats) connect4TupleUpdateFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.connect4TupleUpdateFailuresN
-}
-
-func (s *runStats) setUDPRingbufReserveFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.udpRingbufReserveFailuresN = n
-}
-
-func (s *runStats) udpRingbufReserveFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.udpRingbufReserveFailuresN
-}
-
-func (s *runStats) setDNSRingbufReserveFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.dnsRingbufReserveFailuresN = n
-}
-
-func (s *runStats) dnsRingbufReserveFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.dnsRingbufReserveFailuresN
-}
-
-func (s *runStats) setConnectRingbufReserveFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.connectRingbufReserveFailuresN = n
-}
-
-func (s *runStats) setHTTPRingbufReserveFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.httpRingbufReserveFailuresN = n
-}
-
-func (s *runStats) setTLSRingbufReserveFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.tlsRingbufReserveFailuresN = n
-}
-
-func (s *runStats) setExecRingbufReserveFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.execRingbufReserveFailuresN = n
-}
-
-func (s *runStats) setForkRingbufReserveFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.forkRingbufReserveFailuresN = n
-}
-
-func (s *runStats) setFSRingbufReserveFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.fsRingbufReserveFailuresN = n
-}
-
-func (s *runStats) connectRingbufReserveFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.connectRingbufReserveFailuresN
-}
-
-func (s *runStats) httpRingbufReserveFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.httpRingbufReserveFailuresN
-}
-
-func (s *runStats) tlsRingbufReserveFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.tlsRingbufReserveFailuresN
-}
-
-func (s *runStats) execRingbufReserveFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.execRingbufReserveFailuresN
-}
-
-func (s *runStats) forkRingbufReserveFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.forkRingbufReserveFailuresN
-}
-
-func (s *runStats) fsRingbufReserveFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.fsRingbufReserveFailuresN
-}
-
-func (s *runStats) setUDPSendmsgMultiIovecObserved(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.udpSendmsgMultiIovecObservedN = n
-}
-
-func (s *runStats) udpSendmsgMultiIovecObserved() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.udpSendmsgMultiIovecObservedN
-}
-
-func (s *runStats) setTLSWritevMultiIovecObserved(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.tlsWritevMultiIovecObservedN = n
-}
-
-func (s *runStats) tlsWritevMultiIovecObserved() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.tlsWritevMultiIovecObservedN
-}
-
-func (s *runStats) setUnobservedEgressSyscalls(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.unobservedEgressSyscallsN = n
-}
-
-func (s *runStats) unobservedEgressSyscalls() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.unobservedEgressSyscallsN
-}
-
-func (s *runStats) setIoUringSetupObserved(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.ioUringSetupObservedN = n
-}
-
-func (s *runStats) ioUringSetupObserved() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.ioUringSetupObservedN
-}
-
-func (s *runStats) setTCPDNSResponsesObserved(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.tcpDNSResponsesObservedN = n
-}
-
-func (s *runStats) tcpDNSResponsesObserved() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.tcpDNSResponsesObservedN
-}
-
-func (s *runStats) setTCPDNSSkippedShortRead(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.tcpDNSSkippedShortReadN = n
-}
-
-func (s *runStats) tcpDNSSkippedShortRead() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.tcpDNSSkippedShortReadN
-}
-
-func (s *runStats) addBPFHeartbeatFailure() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.bpfHeartbeatFailures++
-}
-
-func (s *runStats) bpfHeartbeatFailureCount() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.bpfHeartbeatFailures
-}
-
-func (s *runStats) addBPFAudit() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.bpfAuditN++
-}
-
-func (s *runStats) setBPFAuditRingbufReserveFailures(n int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.bpfAuditRingbufReserveFailuresN = n
-}
-
-func (s *runStats) bpfAuditTotal() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.bpfAuditN
-}
-
-func (s *runStats) addBPFMapIntegrityFailure() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.bpfMapIntegrityFailuresN++
-}
-
-func (s *runStats) bpfMapIntegrityFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.bpfMapIntegrityFailuresN
-}
-
-// addDNSCacheUpdateFailure bumps the per-run counter for failed BPF
-// dns_cache map mutations (Update or non-ErrKeyNotExist Delete). Wired to
-// DNSCache.SetBPFFailureCallback so partial sync between userspace and the
-// kernel-side dns_cache instances is observable in the digest (M-09).
-func (s *runStats) addDNSCacheUpdateFailure() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.bpfDNSCacheUpdateFailuresN++
-}
-
-func (s *runStats) bpfAuditRingbufReserveFailures() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.bpfAuditRingbufReserveFailuresN
-}
-
-func (s *runStats) snapshotSummary(kernel string, bpf []telemetry.BPFStatus) telemetry.Summary {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	pc := make(map[string]int, len(s.policyCounts))
-	for k, v := range s.policyCounts {
-		pc[k] = v
-	}
-	dropped := make(map[string]int, len(s.droppedCounts))
-	for k, v := range s.droppedCounts {
-		dropped[k] = v
-	}
-	rbTotal := telemetry.SumRingbufReserveFailuresDetectPath(
-		s.udpRingbufReserveFailuresN,
-		s.dnsRingbufReserveFailuresN,
-		s.connectRingbufReserveFailuresN,
-		s.httpRingbufReserveFailuresN,
-		s.tlsRingbufReserveFailuresN,
-		s.execRingbufReserveFailuresN,
-		s.forkRingbufReserveFailuresN,
-		s.fsRingbufReserveFailuresN,
-		s.bpfAuditRingbufReserveFailuresN,
-	)
-	return telemetry.Summary{
-		Version:                        2,
-		SchemaVersion:                  telemetry.SchemaVersion,
-		ExecEvents:                     s.execN,
-		TCPEvents:                      s.tcpN,
-		UDPEvents:                      s.udpN,
-		HTTPEvents:                     s.httpN,
-		TLSEvents:                      s.tlsN,
-		ProcForkEvents:                 s.procForkN,
-		Connect4TupleUpdateFailures:    s.connect4TupleUpdateFailuresN,
-		UDPRingbufReserveFailures:      s.udpRingbufReserveFailuresN,
-		DNSRingbufReserveFailures:      s.dnsRingbufReserveFailuresN,
-		ConnectRingbufReserveFailures:  s.connectRingbufReserveFailuresN,
-		HTTPRingbufReserveFailures:     s.httpRingbufReserveFailuresN,
-		TLSRingbufReserveFailures:      s.tlsRingbufReserveFailuresN,
-		ExecRingbufReserveFailures:     s.execRingbufReserveFailuresN,
-		ForkRingbufReserveFailures:     s.forkRingbufReserveFailuresN,
-		FSRingbufReserveFailures:       s.fsRingbufReserveFailuresN,
-		RingbufReserveFailuresTotal:    rbTotal,
-		UDPSendmsgMultiIovecObserved:   s.udpSendmsgMultiIovecObservedN,
-		TLSWritevMultiIovecObserved:    s.tlsWritevMultiIovecObservedN,
-		UnobservedEgressSyscalls:       s.unobservedEgressSyscallsN,
-		IoUringSetupObserved:           s.ioUringSetupObservedN,
-		TCPDNSResponsesObserved:        s.tcpDNSResponsesObservedN,
-		TCPDNSSkippedShortRead:         s.tcpDNSSkippedShortReadN,
-		BPFAuditEvents:                 s.bpfAuditN,
-		BPFHeartbeatFailures:           s.bpfHeartbeatFailures,
-		BPFMapIntegrityFailures:        s.bpfMapIntegrityFailuresN,
-		BPFDNSCacheUpdateFailures:      s.bpfDNSCacheUpdateFailuresN,
-		BPFAuditRingbufReserveFailures: s.bpfAuditRingbufReserveFailuresN,
-		DroppedCounts:                  dropped,
-		PolicyCounts:                   pc,
-		KernelRelease:                  kernel,
-		BPF:                            bpf,
-	}
-}
-
-func (s *runStats) counts() (execN, tcpN, udpN, httpN, tlsN, fsN int) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.execN, s.tcpN, s.udpN, s.httpN, s.tlsN, s.fsN
-}
-
-// snapshotPolicyCounts returns a copy of policy classification counters for digests.
-func (s *runStats) snapshotPolicyCounts() map[string]int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	pc := make(map[string]int, len(s.policyCounts))
-	for k, v := range s.policyCounts {
-		pc[k] = v
-	}
-	return pc
-}
-
-type rowBuffer struct {
-	mu   sync.Mutex
-	max  int
-	exec []report.ExecDigestRow
-	tcp  []report.TCPDigestRow
-	udp  []report.UDPDigestRow
-	http []report.HTTPDigestRow
-	tls  []report.TLSDigestRow
-}
-
-func newRowBuffer(max int) *rowBuffer {
-	return &rowBuffer{max: max}
-}
-
-// trimRing trims s to at most max entries (drops oldest); returns the number of dropped entries
-// so callers can record the drop in stats (e.g. runStats.addDropped("<kind>_ring_trim")).
-func trimRing[T any](s *[]T, max int) int {
-	if max <= 0 || len(*s) <= max {
-		return 0
-	}
-	droppedN := len(*s) - max
-	*s = (*s)[droppedN:]
-	slog.Debug("telemetry row buffer trimmed (ring full)", "dropped", droppedN, "retained", max)
-	return droppedN
-}
-
-func (b *rowBuffer) addExec(r report.ExecDigestRow, stats *runStats) {
-	b.mu.Lock()
-	b.exec = append(b.exec, r)
-	dropped := trimRing(&b.exec, b.max)
-	b.mu.Unlock()
-	if dropped > 0 && stats != nil {
-		for i := 0; i < dropped; i++ {
-			stats.addDropped("exec_ring_trim")
-		}
-	}
-}
-
-func (b *rowBuffer) addTCP(r report.TCPDigestRow, stats *runStats) {
-	b.mu.Lock()
-	b.tcp = append(b.tcp, r)
-	dropped := trimRing(&b.tcp, b.max)
-	b.mu.Unlock()
-	if dropped > 0 && stats != nil {
-		for i := 0; i < dropped; i++ {
-			stats.addDropped("tcp_ring_trim")
-		}
-	}
-}
-
-func (b *rowBuffer) addUDP(r report.UDPDigestRow, stats *runStats) {
-	b.mu.Lock()
-	b.udp = append(b.udp, r)
-	dropped := trimRing(&b.udp, b.max)
-	b.mu.Unlock()
-	if dropped > 0 && stats != nil {
-		for i := 0; i < dropped; i++ {
-			stats.addDropped("udp_ring_trim")
-		}
-	}
-}
-
-func (b *rowBuffer) addHTTP(r report.HTTPDigestRow, stats *runStats) {
-	b.mu.Lock()
-	b.http = append(b.http, r)
-	dropped := trimRing(&b.http, b.max)
-	b.mu.Unlock()
-	if dropped > 0 && stats != nil {
-		for i := 0; i < dropped; i++ {
-			stats.addDropped("http_ring_trim")
-		}
-	}
-}
-
-func (b *rowBuffer) addTLS(r report.TLSDigestRow, stats *runStats) {
-	b.mu.Lock()
-	b.tls = append(b.tls, r)
-	dropped := trimRing(&b.tls, b.max)
-	b.mu.Unlock()
-	if dropped > 0 && stats != nil {
-		for i := 0; i < dropped; i++ {
-			stats.addDropped("tls_ring_trim")
-		}
-	}
-}
-
-func (b *rowBuffer) snapshot() (exec []report.ExecDigestRow, tcp []report.TCPDigestRow, udp []report.UDPDigestRow, http []report.HTTPDigestRow, tls []report.TLSDigestRow) {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	return slices.Clone(b.exec), slices.Clone(b.tcp), slices.Clone(b.udp), slices.Clone(b.http), slices.Clone(b.tls)
-}
-
-func setupLogging(level string) {
-	lvl := slog.LevelInfo
-	switch strings.ToLower(strings.TrimSpace(level)) {
-	case "debug":
-		lvl = slog.LevelDebug
-	case "warn", "warning":
-		lvl = slog.LevelWarn
-	case "error":
-		lvl = slog.LevelError
-	}
-	h := slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{Level: lvl})
-	slog.SetDefault(slog.New(h))
-}
-
-func writeAgentStatus(path string, ok bool) error {
-	if path == "" {
-		return nil
-	}
-	dir := filepath.Dir(path)
-	if dir != "" && dir != "." {
-		// #nosec G301 -- directory must be traversable by the non-root runner user polling readiness.
-		if err := os.MkdirAll(dir, 0o755); err != nil {
-			return err
-		}
-	}
-	payload := map[string]any{"ok": ok, "version": 1}
-	b, err := json.Marshal(payload)
-	if err != nil {
-		return err
-	}
-	// GitHub Actions polls this path as the runner user while the agent runs under sudo; 0o600
-	// root-owned files are unreadable (EACCES). Payload is non-secret (ok + version only).
-	// #nosec G306 -- readiness file intentionally world-readable for runner polling semantics.
-	if err := atomicwrite.Bytes(path, b, 0o644); err != nil {
-		return err
-	}
-	slog.Info("agent ready status written", "component", "ready", "path", path, "ok", ok)
-	return nil
-}
-
-func agentVersionString() string {
-	bi, ok := debug.ReadBuildInfo()
-	if !ok {
-		return "unknown"
-	}
-	v := strings.TrimSpace(bi.Main.Version)
-	if v == "" || v == "(devel)" {
-		return "devel"
-	}
-	return v
-}
-
-func kernelRelease() string {
-	var uts unix.Utsname
-	if err := unix.Uname(&uts); err != nil {
-		return ""
-	}
-	return unix.ByteSliceToString(uts.Release[:])
-}
-
-// decodeConnectEvent parses trace_connect.bpf.c connect_event (tgid, tid, comm, daddr, dport).
-func decodeConnectEvent(raw []byte) (tgid, tid uint32, comm [16]byte, daddr [4]byte, dport uint16, ok bool) {
-	if len(raw) < connectEventWireSize {
-		return 0, 0, [16]byte{}, [4]byte{}, 0, false
-	}
-	tgid = binary.LittleEndian.Uint32(raw[0:4])
-	tid = binary.LittleEndian.Uint32(raw[4:8])
-	copy(comm[:], raw[8:24])
-	copy(daddr[:], raw[24:28])
-	dport = binary.BigEndian.Uint16(raw[28:30])
-	return tgid, tid, comm, daddr, dport, true
-}
-
-// decodeUDPSendEvent parses udp_send_event. datagram_len lives at offset 32
-// because of the explicit `__u8 _pad[2]` (offsets 30..32) added in
-// trace_connect_obs.h. Reading from offset 30 like the prior implementation
-// did yields garbage bytes from the implicit alignment pad.
-func decodeUDPSendEvent(raw []byte) (tgid, tid uint32, comm [16]byte, daddr [4]byte, dport uint16, dgramLen uint32, ok bool) {
-	if len(raw) < udpSendEventWireSize {
-		return 0, 0, [16]byte{}, [4]byte{}, 0, 0, false
-	}
-	tgid = binary.LittleEndian.Uint32(raw[0:4])
-	tid = binary.LittleEndian.Uint32(raw[4:8])
-	copy(comm[:], raw[8:24])
-	copy(daddr[:], raw[24:28])
-	dport = binary.BigEndian.Uint16(raw[28:30])
-	dgramLen = binary.LittleEndian.Uint32(raw[32:36])
-	return tgid, tid, comm, daddr, dport, dgramLen, true
-}
-
-const httpPayloadMax = 192
-
-// decodeHTTPSniffEvent parses http_sniff_event (228 bytes with HTTP_PAYLOAD_MAX=192).
-func decodeHTTPSniffEvent(raw []byte) (tgid, tid uint32, comm [16]byte, daddr [4]byte, dport uint16, payload []byte, ok bool) {
-	if len(raw) < httpSniffEventWireSize {
-		return 0, 0, [16]byte{}, [4]byte{}, 0, nil, false
-	}
-	tgid = binary.LittleEndian.Uint32(raw[0:4])
-	tid = binary.LittleEndian.Uint32(raw[4:8])
-	copy(comm[:], raw[8:24])
-	copy(daddr[:], raw[24:28])
-	dport = binary.BigEndian.Uint16(raw[28:30])
-	capLen := int(binary.LittleEndian.Uint16(raw[32:34]))
-	// capLen is derived from Uint16 and cast to int on a 64-bit system;
-	// it is always in [0, 65535]. Only the upper bound needs checking.
-	if capLen > httpPayloadMax {
-		return 0, 0, [16]byte{}, [4]byte{}, 0, nil, false
-	}
-	payload = make([]byte, capLen)
-	copy(payload, raw[httpSniffEventHeaderSize:httpSniffEventHeaderSize+capLen])
-	return tgid, tid, comm, daddr, dport, payload, true
-}
-
-const tlsPayloadMax = 256
-
-// decodeTLSSniffEvent parses tls_sniff_event (same wire layout as http_sniff_event with TLS_PAYLOAD_MAX).
-func decodeTLSSniffEvent(raw []byte) (tgid, tid uint32, comm [16]byte, daddr [4]byte, dport uint16, payload []byte, ok bool) {
-	if len(raw) < tlsSniffEventWireSize {
-		return 0, 0, [16]byte{}, [4]byte{}, 0, nil, false
-	}
-	tgid = binary.LittleEndian.Uint32(raw[0:4])
-	tid = binary.LittleEndian.Uint32(raw[4:8])
-	copy(comm[:], raw[8:24])
-	copy(daddr[:], raw[24:28])
-	dport = binary.BigEndian.Uint16(raw[28:30])
-	capLen := int(binary.LittleEndian.Uint16(raw[32:34]))
-	// capLen is derived from Uint16 and cast to int on a 64-bit system;
-	// it is always in [0, 65535]. Only the upper bound needs checking.
-	if capLen > tlsPayloadMax {
-		return 0, 0, [16]byte{}, [4]byte{}, 0, nil, false
-	}
-	payload = make([]byte, capLen)
-	copy(payload, raw[tlsSniffEventHeaderSize:tlsSniffEventHeaderSize+capLen])
-	return tgid, tid, comm, daddr, dport, payload, true
-}
-
-// decodeDNSSniffSample parses dns_sniff_event (trace_dns.bpf.c): len, is_tcp, pad, payload.
-// Legacy ringbuf records are dnsSniffEventWireSizeLegacy (len + data only).
-func decodeDNSSniffSample(raw []byte) (pkt []byte, isTCP bool, ok bool) {
-	if len(raw) < 4 {
-		return nil, false, false
-	}
-	n := binary.LittleEndian.Uint32(raw[0:4])
-	if n > dnsSniffMaxPayload {
-		return nil, false, false
-	}
-	if len(raw) == dnsSniffEventWireSizeLegacy {
-		if int(n)+4 > len(raw) {
-			return nil, false, false
-		}
-		return raw[4 : 4+int(n)], false, true
-	}
-	if len(raw) < dnsSniffEventWireSize {
-		return nil, false, false
-	}
-	isTCP = raw[4] != 0
-	if int(n)+8 > len(raw) {
-		return nil, false, false
-	}
-	return raw[8 : 8+int(n)], isTCP, true
-}
-
-// decodeBPFAuditEvent parses trace_bpf_audit.bpf.c bpf_audit_event (tgid, tid, cmd, comm).
-// BPF struct layout: tgid(0-3) tid(4-7) cmd(8-11) comm[16](12-27).
-func decodeBPFAuditEvent(raw []byte) (tgid, tid uint32, comm [16]byte, cmd uint32, ok bool) {
-	if len(raw) < bpfAuditEventWireSize {
-		return 0, 0, [16]byte{}, 0, false
-	}
-	tgid = binary.LittleEndian.Uint32(raw[0:4])
-	tid = binary.LittleEndian.Uint32(raw[4:8])
-	cmd = binary.LittleEndian.Uint32(raw[8:12])
-	copy(comm[:], raw[12:28])
-	return tgid, tid, comm, cmd, true
-}
-
-func decodeDenyEvent(raw []byte) (tgid, tid uint32, comm [16]byte, protocol uint8, reason uint8, af uint8,
-	daddr16 [16]byte, dport uint16, ok bool) {
-	if len(raw) < denyEventWireSize {
-		return 0, 0, [16]byte{}, 0, 0, 0, [16]byte{}, 0, false
-	}
-	tgid = binary.LittleEndian.Uint32(raw[0:4])
-	tid = binary.LittleEndian.Uint32(raw[4:8])
-	copy(comm[:], raw[8:24])
-	protocol = raw[24]
-	reason = raw[25]
-	af = raw[26]
-	copy(daddr16[:], raw[28:44])
-	dport = binary.BigEndian.Uint16(raw[44:46])
-	return tgid, tid, comm, protocol, reason, af, daddr16, dport, true
-}
-
-func denyProtocolLabel(protocol uint8) string {
-	switch protocol {
-	case denyProtoTCP:
-		return "tcp"
-	case denyProtoUDP:
-		return "udp"
-	default:
-		return "unknown"
-	}
-}
-
-func denyReasonLabel(reason uint8) string {
-	switch reason {
-	case denyReasonDstNotAllowlisted:
-		return "dst_not_allowlisted"
-	default:
-		return "unknown"
-	}
-}
-
-func denyDigestRowFromEvent(ev telemetry.DenyEvent) report.DenyDigestRow {
-	return report.DenyDigestRow{
-		TS:       ev.TS,
-		PID:      ev.PID,
-		Comm:     ev.Comm,
-		Protocol: ev.Protocol,
-		Dst:      ev.Dst,
-		Dport:    ev.Dport,
-		Reason:   ev.Reason,
-	}
-}
-
-// startSyscallTrace loads observability-only BPF (TCP connect + UDP sendto + HTTP sniff + TLS write sniff; single raw_tp attach).
-// cgroup enforcement loads separately (traceenforce) when mode is enforce.
-// When enableTLSSNI is true, sets tls_agent_cfg map so BPF emits TLS ClientHello captures.
-// tlsAgentCfgFailed is set when the map update fails (SNI path stays off in BPF) so callers can mark the hook degraded.
-func startSyscallTrace(enableTLSSNI bool) (connRd, udpRd, httpRd, tlsRd *ringbuf.Reader, objs *traceconnect.TraceconnectObjects, lnk link.Link, tlsAgentCfgFailed bool, err error) {
-	objs = new(traceconnect.TraceconnectObjects)
-	// Default: fast verifier path (nil opts). Branch + instruction verifier logging makes
-	// LoadTraceconnectObjects disproportionately slow on hosted runners and can exceed the
-	// composite action's waitForAgentReady window (see src/main.ts). Opt in via env for debugging:
-	//   COLDSTEP_BPF_VERBOSE_VERIFY=1
-	var traceLoadOpts *ebpf.CollectionOptions
-	if strings.TrimSpace(os.Getenv("COLDSTEP_BPF_VERBOSE_VERIFY")) != "" {
-		traceLoadOpts = &ebpf.CollectionOptions{
-			Programs: ebpf.ProgramOptions{
-				LogLevel:     ebpf.LogLevelBranch | ebpf.LogLevelInstruction,
-				LogSizeStart: 512 * 1024,
-			},
-		}
-	}
-	if err = traceconnect.LoadTraceconnectObjects(objs, traceLoadOpts); err != nil {
-		return nil, nil, nil, nil, nil, nil, false, err
-	}
-
-	lnk, err = link.AttachRawTracepoint(link.RawTracepointOptions{
-		Name:    "sys_enter",
-		Program: objs.HandleRawSysEnter,
-	})
-	if err != nil {
-		_ = objs.Close()
-		return nil, nil, nil, nil, nil, nil, false, err
-	}
-
-	connRd, err = ringbuf.NewReader(objs.ConnectEvents)
-	if err != nil {
-		_ = lnk.Close()
-		_ = objs.Close()
-		return nil, nil, nil, nil, nil, nil, false, err
-	}
-	udpRd, err = ringbuf.NewReader(objs.UdpEvents)
-	if err != nil {
-		_ = connRd.Close()
-		_ = lnk.Close()
-		_ = objs.Close()
-		return nil, nil, nil, nil, nil, nil, false, err
-	}
-	httpRd, err = ringbuf.NewReader(objs.HttpEvents)
-	if err != nil {
-		_ = udpRd.Close()
-		_ = connRd.Close()
-		_ = lnk.Close()
-		_ = objs.Close()
-		return nil, nil, nil, nil, nil, nil, false, err
-	}
-	tlsRd, err = ringbuf.NewReader(objs.TlsEvents)
-	if err != nil {
-		_ = httpRd.Close()
-		_ = udpRd.Close()
-		_ = connRd.Close()
-		_ = lnk.Close()
-		_ = objs.Close()
-		return nil, nil, nil, nil, nil, nil, false, err
-	}
-
-	if enableTLSSNI {
-		if uerr := objs.TlsAgentCfg.Update(uint32(0), uint8(1), ebpf.UpdateAny); uerr != nil {
-			tlsAgentCfgFailed = true
-			slog.Warn("tls_sni bpf cfg", "err", uerr)
-		}
-	}
-
-	return connRd, udpRd, httpRd, tlsRd, objs, lnk, tlsAgentCfgFailed, nil
-}
-
-func startBPFAuditTrace() (rd *ringbuf.Reader, objs *tracebpfaudit.TracebpfauditObjects, lnk link.Link, err error) {
-	objs = new(tracebpfaudit.TracebpfauditObjects)
-	if err = tracebpfaudit.LoadTracebpfauditObjects(objs, nil); err != nil {
-		return nil, nil, nil, err
-	}
-
-	lnk, err = link.AttachRawTracepoint(link.RawTracepointOptions{
-		Name:    "sys_enter",
-		Program: objs.HandleRawSysEnterBpf,
-	})
-	if err != nil {
-		_ = objs.Close()
-		return nil, nil, nil, err
-	}
-
-	rd, err = ringbuf.NewReader(objs.BpfAuditEvents)
-	if err != nil {
-		_ = lnk.Close()
-		_ = objs.Close()
-		return nil, nil, nil, err
-	}
-	return rd, objs, lnk, nil
-}
-
-func startDNSTrace() (*ringbuf.Reader, *tracedns.TracednsObjects, link.Link, link.Link, error) {
-	objs := new(tracedns.TracednsObjects)
-	if err := tracedns.LoadTracednsObjects(objs, nil); err != nil {
-		return nil, nil, nil, nil, err
-	}
-
-	lnkEnter, err := link.AttachRawTracepoint(link.RawTracepointOptions{
-		Name:    "sys_enter",
-		Program: objs.HandleRawSysEnterDns,
-	})
-	if err != nil {
-		_ = objs.Close()
-		return nil, nil, nil, nil, err
-	}
-
-	lnkExit, err := link.AttachRawTracepoint(link.RawTracepointOptions{
-		Name:    "sys_exit",
-		Program: objs.HandleRawSysExitDns,
-	})
-	if err != nil {
-		_ = lnkEnter.Close()
-		_ = objs.Close()
-		return nil, nil, nil, nil, err
-	}
-
-	rd, err := ringbuf.NewReader(objs.DnsEvents)
-	if err != nil {
-		_ = lnkExit.Close()
-		_ = lnkEnter.Close()
-		_ = objs.Close()
-		return nil, nil, nil, nil, err
-	}
-
-	return rd, objs, lnkEnter, lnkExit, nil
-}
-
-func readExecRing(ctx context.Context, cfg config.Config, rd *ringbuf.Reader, stats *runStats,
-	rows *rowBuffer, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, signer *telemetry.Signer) error {
-	backoff := newRingReadRetryBackoff()
-	for {
-		record, err := rd.Read()
-		if err != nil {
-			if errors.Is(err, ringbuf.ErrClosed) {
-				return nil
-			}
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			delay := backoff.sleep()
-			slog.Warn("ringbuf read (exec)", "err", err, "backoff", delay)
-			continue
-		}
-		backoff.reset()
-
-		var ev execEvent
-		if err := binary.Read(bytes.NewReader(record.RawSample), binary.LittleEndian, &ev); err != nil {
-			stats.addDropped("exec_decode")
-			slog.Warn("decode exec", "err", err)
-			continue
-		}
-
-		comm := string(bytes.TrimRight(ev.Comm[:], "\x00"))
-		exe := string(bytes.TrimRight(ev.ExePath[:], "\x00"))
-		stats.addExec()
-		ts := time.Now().UTC().Format(time.RFC3339Nano)
-		rows.addExec(report.ExecDigestRow{
-			TS: ts, PID: ev.TGID, ThreadID: ev.TID, Comm: comm,
-			Exe: report.TruncateExeForDigest(exe),
-		}, stats)
-
-		if cfg.EventsLogPath != "" {
-			jsonlMu.Lock()
-			n := seq.Next()
-			evOut := telemetry.ExecEvent{
-				Type: "exec", TS: ts, Seq: n,
-				PID: ev.TGID, TGID: ev.TGID, ThreadID: ev.TID, Comm: comm,
-				Exe: exe,
-			}
-			err := telemetry.AppendJSONL(cfg.EventsLogPath, evOut, signer)
-			jsonlMu.Unlock()
-			if err != nil {
-				stats.addDropped("exec_jsonl")
-				slog.Warn("events jsonl", "err", err)
-			}
-		}
-	}
-}
-
-type forkEventWire struct {
-	ParentPID     uint32
-	ChildPID      uint32
-	ParentComm    [16]byte
-	ChildComm     [16]byte
-	ChildSID      uint32 // v0.3: session leader PID
-	ChildPidnsNum uint32 // v0.3: PID namespace inode number
-}
-
-func readForkRing(ctx context.Context, cfg config.Config, rd *ringbuf.Reader, stats *runStats,
-	forkBuf *forkEdgeBuffer, forkState *forkSectionState, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, signer *telemetry.Signer) error {
-	backoff := newRingReadRetryBackoff()
-	for {
-		record, err := rd.Read()
-		if err != nil {
-			if errors.Is(err, ringbuf.ErrClosed) {
-				return nil
-			}
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			if forkState != nil {
-				forkState.addReadError()
-			}
-			delay := backoff.sleep()
-			slog.Warn("ringbuf read (fork)", "err", err, "backoff", delay)
-			continue
-		}
-		backoff.reset()
-
-		var ev forkEventWire
-		if err := binary.Read(bytes.NewReader(record.RawSample), binary.LittleEndian, &ev); err != nil {
-			forkState.addReadError()
-			stats.addDropped("proc_fork_decode")
-			slog.Warn("decode fork", "err", err)
-			continue
-		}
-
-		pcomm := string(bytes.TrimRight(ev.ParentComm[:], "\x00"))
-		ccomm := string(bytes.TrimRight(ev.ChildComm[:], "\x00"))
-		forkBuf.add(proctree.Edge{
-			ParentTGID:    ev.ParentPID,
-			ChildTGID:     ev.ChildPID,
-			ParentComm:    pcomm,
-			ChildComm:     ccomm,
-			ChildSID:      ev.ChildSID,
-			ChildPidnsNum: ev.ChildPidnsNum,
-		})
-		stats.addProcFork()
-		ts := time.Now().UTC().Format(time.RFC3339Nano)
-
-		if cfg.EventsLogPath != "" {
-			jsonlMu.Lock()
-			n := seq.Next()
-			evOut := telemetry.ProcForkEvent{
-				Type: "proc_fork", TS: ts, Seq: n,
-				ParentPID: ev.ParentPID, ChildPID: ev.ChildPID,
-				ParentComm: pcomm, ChildComm: ccomm,
-				ChildSID:      ev.ChildSID,
-				ChildPidnsNum: ev.ChildPidnsNum,
-				Note:          "best-effort pid namespace; parent/child are kernel fork trace ids",
-			}
-			werr := telemetry.AppendJSONL(cfg.EventsLogPath, evOut, signer)
-			jsonlMu.Unlock()
-			if werr != nil {
-				stats.addDropped("proc_fork_jsonl")
-				slog.Warn("events jsonl", "err", werr)
-			}
-		}
-	}
-}
-
-func nullTermStr(b []byte) string {
-	return string(bytes.TrimRight(b, "\x00"))
-}
-
-// fsOpName maps BPF op byte to JSONL op string.
-func fsOpName(op uint8) string {
-	switch op {
-	case 1:
-		return "create"
-	case 2:
-		return "unlink"
-	case 3:
-		return "rename"
-	case 4:
-		return "chmod"
-	default:
-		return "unknown"
-	}
-}
-
-type fsEventWire struct {
-	TGID uint32
-	TID  uint32
-	Comm [16]byte
-	Op   uint8
-	Path [256]byte
-	Pad  [3]byte
-}
-
-const maxFSEventsTotal = 5000
-
-func readFSRing(ctx context.Context, cfg config.Config, rd *ringbuf.Reader, stats *runStats,
-	fsRows *fsRowBuffer, fsState *fsSectionState, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, signer *telemetry.Signer) error {
-	count := 0
-	backoff := newRingReadRetryBackoff()
-	for {
-		rec, err := rd.Read()
-		if err != nil {
-			if errors.Is(err, ringbuf.ErrClosed) {
-				return nil
-			}
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			fsState.addReadError()
-			delay := backoff.sleep()
-			slog.Warn("ringbuf read (fs)", "err", err, "backoff", delay)
-			continue
-		}
-		backoff.reset()
-		var ev fsEventWire
-		if err := binary.Read(bytes.NewReader(rec.RawSample), binary.LittleEndian, &ev); err != nil {
-			fsState.addReadError()
-			stats.addDropped("fs_decode")
-			slog.Warn("decode fs event", "err", err)
-			continue
-		}
-
-		count++
-		stats.addFS() // count all events even when rate-capped
-		if count > maxFSEventsTotal {
-			stats.addDropped("fs_cap")
-			if count == maxFSEventsTotal+1 {
-				slog.Warn("fs event cap reached; further events counted but not written to JSONL or rows", "cap", maxFSEventsTotal)
-			}
-			continue
-		}
-
-		comm := nullTermStr(ev.Comm[:])
-		path := nullTermStr(ev.Path[:])
-		op := fsOpName(ev.Op)
-		ts := time.Now().UTC().Format(time.RFC3339Nano)
-
-		fsRows.add(report.FSDigestRow{
-			TS:   ts,
-			PID:  ev.TGID,
-			Comm: comm,
-			Op:   op,
-			Path: path,
-		})
-
-		if cfg.EventsLogPath != "" {
-			jsonlMu.Lock()
-			n := seq.Next()
-			evOut := telemetry.FSEvent{
-				Type: "fs_event", TS: ts, Seq: n,
-				PID: ev.TGID, TGID: ev.TGID, ThreadID: ev.TID,
-				Comm: comm, Op: op, Path: path,
-			}
-			werr := telemetry.AppendJSONL(cfg.EventsLogPath, evOut, signer)
-			jsonlMu.Unlock()
-			if werr != nil {
-				stats.addDropped("fs_jsonl")
-				slog.Warn("events jsonl", "err", werr)
-			}
-		}
-	}
-}
-
-func readConnectRing(ctx context.Context, cfg config.Config, rd *ringbuf.Reader, dns *DNSCache,
-	pol *policy.Policy, stats *runStats, rows *rowBuffer, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, sectionState *networkSectionState, canary *canaryState, signer *telemetry.Signer) error {
-	backoff := newRingReadRetryBackoff()
-	for {
-		record, err := rd.Read()
-		if err != nil {
-			if errors.Is(err, ringbuf.ErrClosed) {
-				return nil
-			}
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			if sectionState != nil {
-				sectionState.addTCPReaderError()
-			}
-			delay := backoff.sleep()
-			slog.Warn("ringbuf read (tcp)", "err", err, "backoff", delay)
-			continue
-		}
-		backoff.reset()
-
-		// Canary event detection: if the record starts with CANARY_MAGIC,
-		// it's a telemetry integrity canary — not a connect event.
-		if len(record.RawSample) >= canaryEventWireSize {
-			magic := binary.LittleEndian.Uint32(record.RawSample[0:4])
-			if magic == canaryMagic {
-				seqNr := binary.LittleEndian.Uint64(record.RawSample[8:16])
-				if canary != nil {
-					canary.noteReceived(seqNr)
-				}
-				slog.Debug("canary received", "seq", seqNr)
-				continue
-			}
-		}
-
-		tgid, tid, commb, daddr, port, decOK := decodeConnectEvent(record.RawSample)
-		if !decOK {
-			if sectionState != nil {
-				sectionState.addTCPDecodeError()
-			}
-			stats.addDropped("tcp_decode")
-			slog.Warn("decode tcp", "len", len(record.RawSample))
-			continue
-		}
-
-		ip := net.IP(daddr[:]).To4()
-		if ip == nil {
-			continue
-		}
-		comm := string(bytes.TrimRight(commb[:], "\x00"))
-		fqdn, fqdnProv := "", "unknown"
-		if dns != nil {
-			fqdn, fqdnProv = dns.LookupProvenance(ip)
-		}
-		cl := pol.Classify(fqdn, ip)
-		stats.addTCP(cl)
-
-		ts := time.Now().UTC().Format(time.RFC3339Nano)
-		notes := "—"
-		if fqdn != "" {
-			notes = fmt.Sprintf("fqdn `%s` (%s)", report.SanitizeForMarkdown(fqdn), fqdnProv)
-		}
-		rows.addTCP(report.TCPDigestRow{
-			TS: ts, PID: tgid, Comm: comm,
-			Remote: fmt.Sprintf("`%s:%d`", ip.String(), port),
-			Notes:  notes,
-			Policy: cl.Display(),
-		}, stats)
-
-		if cfg.EventsLogPath != "" {
-			jsonlMu.Lock()
-			n := seq.Next()
-			ev := telemetry.TCPEvent{
-				Type: "tcp", TS: ts, Seq: n,
-				PID: tgid, TGID: tgid, ThreadID: tid,
-				Comm: comm, Dst: ip.String(), Dport: port,
-				FQDN: fqdn, FQDNProvenance: fqdnProv,
-				Direction: "egress",
-				Policy:    string(cl),
-			}
-			err := telemetry.AppendJSONL(cfg.EventsLogPath, ev, signer)
-			jsonlMu.Unlock()
-			if err != nil {
-				stats.addDropped("tcp_jsonl")
-				slog.Warn("events jsonl", "err", err)
-			}
-		}
-
-		slog.Debug("tcp", "tgid", tgid, "comm", comm, "dst", ip.String(), "dport", port, "policy", string(cl))
-	}
-}
-
-func readTLSRing(ctx context.Context, cfg config.Config, rd *ringbuf.Reader, pol *policy.Policy,
-	stats *runStats, rows *rowBuffer, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, sectionState *networkSectionState, signer *telemetry.Signer) error {
-	backoff := newRingReadRetryBackoff()
-	for {
-		record, err := rd.Read()
-		if err != nil {
-			if errors.Is(err, ringbuf.ErrClosed) {
-				return nil
-			}
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			if sectionState != nil {
-				sectionState.addTLSReaderError()
-			}
-			delay := backoff.sleep()
-			slog.Warn("ringbuf read (tls)", "err", err, "backoff", delay)
-			continue
-		}
-		backoff.reset()
-
-		tgid, tid, commb, daddr, port, rawPay, ok := decodeTLSSniffEvent(record.RawSample)
-		if !ok {
-			if sectionState != nil {
-				sectionState.addTLSDecodeError()
-			}
-			stats.addDropped("tls_decode")
-			slog.Warn("decode tls sniff", "len", len(record.RawSample))
-			continue
-		}
-		ip := net.IP(daddr[:]).To4()
-		if ip == nil {
-			continue
-		}
-		comm := string(bytes.TrimRight(commb[:], "\x00"))
-		sni, parsed := telemetry.ParseClientHelloSNI(rawPay)
-		if !parsed {
-			stats.addDropped("tls_sni_parse")
-			continue
-		}
-		cl := pol.Classify(sni, ip)
-		stats.addTLS(cl)
-
-		ts := time.Now().UTC().Format(time.RFC3339Nano)
-		rows.addTLS(report.TLSDigestRow{
-			TS: ts, PID: tgid, Comm: comm,
-			SNI:    sni,
-			Remote: fmt.Sprintf("`%s:%d`", ip.String(), port),
-			Policy: cl.Display(),
-		}, stats)
-
-		if cfg.EventsLogPath != "" {
-			jsonlMu.Lock()
-			n := seq.Next()
-			ev := telemetry.TLSEvent{
-				Type: "tls", TS: ts, Seq: n,
-				PID: tgid, TGID: tgid, ThreadID: tid,
-				Comm: comm, SNI: sni,
-				Dst: ip.String(), Dport: port,
-				Policy: string(cl),
-				Note:   "ClientHello SNI from first write(2) buffer; fragmented handshakes may be missed",
-			}
-			err := telemetry.AppendJSONL(cfg.EventsLogPath, ev, signer)
-			jsonlMu.Unlock()
-			if err != nil {
-				stats.addDropped("tls_jsonl")
-				slog.Warn("events jsonl", "err", err)
-			}
-		}
-	}
-}
-
-func readUDPRing(ctx context.Context, cfg config.Config, rd *ringbuf.Reader, dns *DNSCache,
-	pol *policy.Policy, stats *runStats, rows *rowBuffer, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, sectionState *networkSectionState, signer *telemetry.Signer) error {
-	backoff := newRingReadRetryBackoff()
-	for {
-		record, err := rd.Read()
-		if err != nil {
-			if errors.Is(err, ringbuf.ErrClosed) {
-				return nil
-			}
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			if sectionState != nil {
-				sectionState.addUDPReaderError()
-			}
-			delay := backoff.sleep()
-			slog.Warn("ringbuf read (udp)", "err", err, "backoff", delay)
-			continue
-		}
-		backoff.reset()
-
-		tgid, tid, commb, daddr, port, dgramLen, ok := decodeUDPSendEvent(record.RawSample)
-		if !ok {
-			if sectionState != nil {
-				sectionState.addUDPDecodeError()
-			}
-			stats.addDropped("udp_decode")
-			slog.Warn("decode udp", "len", len(record.RawSample))
-			continue
-		}
-		ip := net.IP(daddr[:]).To4()
-		if ip == nil {
-			continue
-		}
-		comm := string(bytes.TrimRight(commb[:], "\x00"))
-		fqdn, fqdnProv := "", "unknown"
-		if dns != nil {
-			fqdn, fqdnProv = dns.LookupProvenance(ip)
-		}
-		cl := pol.Classify(fqdn, ip)
-		stats.addUDP(cl)
-
-		ts := time.Now().UTC().Format(time.RFC3339Nano)
-		rows.addUDP(report.UDPDigestRow{
-			TS: ts, PID: tgid, Comm: comm,
-			Remote:   fmt.Sprintf("`%s:%d`", ip.String(), port),
-			DgramLen: dgramLen,
-			FQDN:     fqdn,
-			Policy:   cl.Display(),
-		}, stats)
-
-		if cfg.EventsLogPath != "" {
-			jsonlMu.Lock()
-			n := seq.Next()
-			ev := telemetry.UDPEvent{
-				Type: "udp", TS: ts, Seq: n,
-				PID: tgid, TGID: tgid, ThreadID: tid,
-				Comm: comm, Dst: ip.String(), Dport: port,
-				DatagramLen: dgramLen, FQDN: fqdn, FQDNProvenance: fqdnProv,
-				Direction: "egress",
-				Policy:    string(cl),
-			}
-			err := telemetry.AppendJSONL(cfg.EventsLogPath, ev, signer)
-			jsonlMu.Unlock()
-			if err != nil {
-				stats.addDropped("udp_jsonl")
-				slog.Warn("events jsonl", "err", err)
-			}
-		}
-	}
-}
-
-func readHTTPRing(ctx context.Context, cfg config.Config, rd *ringbuf.Reader, pol *policy.Policy,
-	stats *runStats, rows *rowBuffer, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, sectionState *networkSectionState, signer *telemetry.Signer) error {
-	backoff := newRingReadRetryBackoff()
-	for {
-		record, err := rd.Read()
-		if err != nil {
-			if errors.Is(err, ringbuf.ErrClosed) {
-				return nil
-			}
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			if sectionState != nil {
-				sectionState.addHTTPReaderError()
-			}
-			delay := backoff.sleep()
-			slog.Warn("ringbuf read (http)", "err", err, "backoff", delay)
-			continue
-		}
-		backoff.reset()
-
-		tgid, tid, commb, daddr, port, rawPay, ok := decodeHTTPSniffEvent(record.RawSample)
-		if !ok {
-			if sectionState != nil {
-				sectionState.addHTTPDecodeError()
-			}
-			stats.addDropped("http_decode")
-			slog.Warn("decode http sniff", "len", len(record.RawSample))
-			continue
-		}
-		ip := net.IP(daddr[:]).To4()
-		if ip == nil {
-			continue
-		}
-		comm := string(bytes.TrimRight(commb[:], "\x00"))
-		method, host, path, parsed := telemetry.ParseHTTPRequestPrefix(rawPay)
-		if !parsed {
-			stats.addDropped("http_prefix_parse")
-			continue
-		}
-		cl := pol.Classify(host, ip)
-		stats.addHTTP(cl)
-
-		ts := time.Now().UTC().Format(time.RFC3339Nano)
-		sumPath := telemetry.RedactPathForSummary(path)
-		rows.addHTTP(report.HTTPDigestRow{
-			TS: ts, PID: tgid, Comm: comm,
-			Method: method, Host: host, Path: sumPath,
-			Remote: fmt.Sprintf("`%s:%d`", ip.String(), port),
-			Policy: cl.Display(),
-		}, stats)
-
-		if cfg.EventsLogPath != "" {
-			jsonlMu.Lock()
-			n := seq.Next()
-			ev := telemetry.HTTPEvent{
-				Type: "http", TS: ts, Seq: n,
-				PID: tgid, TGID: tgid, ThreadID: tid,
-				Comm: comm, Method: method, Host: host, Path: sumPath,
-				Dst: ip.String(), Dport: port,
-				Policy: string(cl),
-			}
-			err := telemetry.AppendJSONL(cfg.EventsLogPath, ev, signer)
-			jsonlMu.Unlock()
-			if err != nil {
-				stats.addDropped("http_jsonl")
-				slog.Warn("events jsonl", "err", err)
-			}
-		}
-	}
-}
-
-func readDNSRing(ctx context.Context, rd *ringbuf.Reader, cache *DNSCache, stats *runStats) error {
-	backoff := newRingReadRetryBackoff()
-	for {
-		record, err := rd.Read()
-		if err != nil {
-			if errors.Is(err, ringbuf.ErrClosed) {
-				return nil
-			}
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			delay := backoff.sleep()
-			slog.Warn("ringbuf read (dns)", "err", err, "backoff", delay)
-			continue
-		}
-		backoff.reset()
-		pkt, isTCP, ok := decodeDNSSniffSample(record.RawSample)
-		if !ok || len(pkt) < 12 {
-			stats.addDropped("dns_decode")
-			continue
-		}
-		if isTCP {
-			// Strip RFC 1035 TCP framing 2-byte length prefix before the DNS header.
-			if len(pkt) < 14 {
-				stats.addDropped("dns_decode_tcp_short")
-				continue
-			}
-			pkt = pkt[2:]
-		}
-		if len(pkt) < 12 {
-			stats.addDropped("dns_decode")
-			continue
-		}
-		cache.AddFromPacket(pkt)
-	}
-}
-
-func readBPFAuditRing(ctx context.Context, cfg config.Config, rd *ringbuf.Reader, stats *runStats, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, signer *telemetry.Signer) error {
-	backoff := newRingReadRetryBackoff()
-	for {
-		record, err := rd.Read()
-		if err != nil {
-			if errors.Is(err, ringbuf.ErrClosed) {
-				return nil
-			}
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			delay := backoff.sleep()
-			slog.Warn("ringbuf read (bpf_audit)", "err", err, "backoff", delay)
-			continue
-		}
-		backoff.reset()
-
-		tgid, tid, commb, cmd, ok := decodeBPFAuditEvent(record.RawSample)
-		if !ok {
-			stats.addDropped("bpf_audit_decode")
-			slog.Warn("decode bpf audit", "len", len(record.RawSample))
-			continue
-		}
-
-		stats.addBPFAudit()
-		comm := string(bytes.TrimRight(commb[:], "\x00"))
-		ts := time.Now().UTC().Format(time.RFC3339Nano)
-
-		slog.Info("bpf syscall audit", "tgid", tgid, "comm", comm, "cmd", cmd)
-
-		if cfg.EventsLogPath != "" {
-			jsonlMu.Lock()
-			n := seq.Next()
-			ev := telemetry.BPFAuditEvent{
-				Type: "bpf_audit", TS: ts, Seq: n,
-				PID: tgid, TGID: tgid, ThreadID: tid,
-				Comm: comm, Cmd: cmd,
-			}
-			err := telemetry.AppendJSONL(cfg.EventsLogPath, ev, signer)
-			jsonlMu.Unlock()
-			if err != nil {
-				stats.addDropped("bpf_audit_jsonl")
-				slog.Warn("events jsonl (bpf_audit)", "err", err)
-			}
-		}
-	}
-}
-
-// integrityBackoffWindow caps how long a single asset's failures are
-// downgraded to slog.Warn after the first slog.Error. Within this window
-// each fresh failure logs as Warn (deduplicated severity); after the window
-// elapses, or after a successful re-arm clears the asset, the next failure
-// re-escalates to Error. JSONL emission and counter increments are unchanged
-// so M-12's bpf_tamper-based hard-fail keeps signaling.
-const integrityBackoffWindow = 5 * time.Minute
-
-// integrityBackoff tracks per-asset last-failure timestamps so recurring
-// integrity failures within integrityBackoffWindow downgrade their slog
-// level (M-13). Each watchMapIntegrity goroutine owns its own instance —
-// the enforce and LSM watchers run independently and reuse asset names.
-type integrityBackoff struct {
-	mu       sync.Mutex
-	lastFail map[string]time.Time
-}
-
-func newIntegrityBackoff() *integrityBackoff {
-	return &integrityBackoff{lastFail: make(map[string]time.Time)}
-}
-
-// noteFailure records a fresh failure for asset and returns true when the
-// caller should escalate to slog.Error (first failure or first failure
-// after the backoff window expired). Subsequent failures inside the window
-// return false so the caller can degrade to slog.Warn.
-func (b *integrityBackoff) noteFailure(asset string) bool {
-	if b == nil {
-		return true
-	}
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	now := time.Now()
-	last, ok := b.lastFail[asset]
-	b.lastFail[asset] = now
-	if !ok {
-		return true
-	}
-	return now.Sub(last) >= integrityBackoffWindow
-}
-
-// clear forgets the backoff state for asset so the next failure re-escalates
-// to slog.Error. Called after a successful revert / re-arm.
-func (b *integrityBackoff) clear(asset string) {
-	if b == nil {
-		return
-	}
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	delete(b.lastFail, asset)
-}
-
-func watchMapIntegrity(ctx context.Context, cfg config.Config, enforceCfg, allowedIpv4, ignoredIpv4 *ebpf.Map, enforceCompiled policy.CompileResult, pol *policy.Policy, stats *runStats, enforceState *enforcementState, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, signer *telemetry.Signer) error {
-	ticker := time.NewTicker(30 * time.Second)
-	defer ticker.Stop()
-
-	backoff := newIntegrityBackoff()
-
-	for {
-		select {
-		case <-ctx.Done():
-			// Shutdown via SIGTERM/SIGINT — avoid treating cancellation like an operational reader failure.
-			if errors.Is(ctx.Err(), context.Canceled) {
-				return nil
-			}
-			return ctx.Err()
-		case <-ticker.C:
-			checkMapIntegrity(cfg, enforceCfg, allowedIpv4, ignoredIpv4, enforceCompiled, pol, stats, enforceState, backoff, seq, jsonlMu, signer)
-		}
-	}
-}
-
-func checkMapIntegrity(cfg config.Config, enforceCfg, allowedIpv4, ignoredIpv4 *ebpf.Map, enforceCompiled policy.CompileResult, pol *policy.Policy, stats *runStats, enforceState *enforcementState, backoff *integrityBackoff, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, signer *telemetry.Signer) {
-	if enforceCfg == nil || allowedIpv4 == nil || ignoredIpv4 == nil {
-		return
-	}
-
-	// 1. Check enforce_cfg
-	const assetEnforceCfg = "map:enforce_cfg"
-	var key uint32 = 0
-	var val uint32
-	modeEnforce := uint32(1)
-	if err := enforceCfg.Lookup(&key, &val); err != nil {
-		logMapIntegrityFailure(cfg, assetEnforceCfg, "lookup error", "", "", stats, seq, jsonlMu, enforceState, backoff, signer)
-		// H-05: a missing/unreadable enforce_cfg key behaves like detect mode in
-		// BPF (`enforcement_enabled()` returns false when the key is absent).
-		// Try to restore the enforce mode key on the same path the value-mismatch
-		// branch uses so a transient lookup failure or tamper does not silently
-		// disable enforcement.
-		if updErr := enforceCfg.Update(&key, &modeEnforce, ebpf.UpdateAny); updErr != nil {
-			slog.Error("BPF map enforce_cfg revert failed (lookup error)", "err", updErr)
-		} else {
-			slog.Error("BPF map enforce_cfg revert succeeded after lookup error")
-			backoff.clear(assetEnforceCfg)
-		}
-	} else if val != 1 {
-		logMapIntegrityFailure(cfg, assetEnforceCfg, "value mismatch", "1", fmt.Sprintf("%d", val), stats, seq, jsonlMu, enforceState, backoff, signer)
-		// Revert tampering.
-		if err := enforceCfg.Update(&key, &modeEnforce, ebpf.UpdateAny); err != nil {
-			slog.Error("BPF map enforce_cfg revert failed", "err", err)
-		} else {
-			backoff.clear(assetEnforceCfg)
-		}
-	}
-
-	// 2. Check allowed_ipv4 count
-	const assetAllowed = "map:allowed_ipv4"
-	count := 0
-	iter := allowedIpv4.Iterate()
-	var k [8]byte // LPM key (4 prefixlen + 4 ip)
-	var v uint8
-	for iter.Next(&k, &v) {
-		count++
-	}
-	if err := iter.Err(); err != nil {
-		logMapIntegrityFailure(cfg, assetAllowed, "iterate error", "", "", stats, seq, jsonlMu, enforceState, backoff, signer)
-	} else {
-		enforceState.mu.Lock()
-		expected := enforceState.expectedEntries
-		enforceState.mu.Unlock()
-		if count != expected {
-			logMapIntegrityFailure(cfg, assetAllowed, "count mismatch", fmt.Sprintf("%d", expected), fmt.Sprintf("%d", count), stats, seq, jsonlMu, enforceState, backoff, signer)
-			// H-04: re-program the LPM trie from the compiled snapshot so a
-			// tampered widening (extra allowed entries) or count corruption
-			// does not persist until process restart.
-			added, removed, rearmErr := rearmAllowedFromSnapshot(allowedIpv4, enforceCompiled, pol)
-			if rearmErr != nil {
-				slog.Error("BPF allowlist re-arm failed", "asset", assetAllowed, "err", rearmErr)
-			} else {
-				slog.Error("BPF allowlist re-armed after tamper", "asset", assetAllowed, "removed", removed, "added", added)
-				backoff.clear(assetAllowed)
-			}
-		}
-	}
-
-	// 3. Check ignored_ipv4 count
-	const assetIgnored = "map:ignored_ipv4_lpm"
-	countIgnored := 0
-	iterIgnored := ignoredIpv4.Iterate()
-	for iterIgnored.Next(&k, &v) {
-		countIgnored++
-	}
-	if err := iterIgnored.Err(); err != nil {
-		logMapIntegrityFailure(cfg, assetIgnored, "iterate error", "", "", stats, seq, jsonlMu, enforceState, backoff, signer)
-	} else {
-		enforceState.mu.Lock()
-		expectedIgnored := enforceState.expectedIgnoredEntries
-		enforceState.mu.Unlock()
-		if countIgnored != expectedIgnored {
-			logMapIntegrityFailure(cfg, assetIgnored, "count mismatch", fmt.Sprintf("%d", expectedIgnored), fmt.Sprintf("%d", countIgnored), stats, seq, jsonlMu, enforceState, backoff, signer)
-			// H-04: same self-heal posture as allowed_ipv4 — restore from
-			// policy.IgnoredIPv4Nets so an attacker cannot widen the
-			// implicit-allow surface by injecting extra ignored CIDRs.
-			added, removed, rearmErr := rearmIgnoredFromPolicy(ignoredIpv4, pol)
-			if rearmErr != nil {
-				slog.Error("BPF allowlist re-arm failed", "asset", assetIgnored, "err", rearmErr)
-			} else {
-				slog.Error("BPF allowlist re-armed after tamper", "asset", assetIgnored, "removed", removed, "added", added)
-				backoff.clear(assetIgnored)
-			}
-		}
-	}
-}
-
-func logMapIntegrityFailure(cfg config.Config, asset, errStr, expected, actual string, stats *runStats, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, enforceState *enforcementState, backoff *integrityBackoff, signer *telemetry.Signer) {
-	stats.addBPFMapIntegrityFailure()
-	if enforceState != nil {
-		enforceState.addMapIntegrityFailure()
-	}
-	// M-13: dedupe slog severity per asset within integrityBackoffWindow. The
-	// JSONL bpf_tamper event and counter increments still flow on every tick
-	// so M-12 (anti-blindness gating) keeps a stable signal — only the
-	// operator-facing log level is dampened.
-	if backoff.noteFailure(asset) {
-		slog.Error("BPF map integrity failure", "asset", asset, "error", errStr, "expected", expected, "actual", actual)
-	} else {
-		slog.Warn("BPF map integrity failure (recurring within backoff window)",
-			"asset", asset, "error", errStr, "expected", expected, "actual", actual,
-			"backoff_window", integrityBackoffWindow)
-	}
-	if cfg.EventsLogPath != "" {
-		jsonlMu.Lock()
-		defer jsonlMu.Unlock()
-		n := seq.Next()
-		ts := time.Now().UTC().Format(time.RFC3339Nano)
-		ev := telemetry.BPFTamperEvent{
-			Type:     "bpf_tamper",
-			TS:       ts,
-			Seq:      n,
-			Asset:    asset,
-			Error:    errStr,
-			Expected: expected,
-			Actual:   actual,
-		}
-		if err := telemetry.AppendJSONL(cfg.EventsLogPath, ev, signer); err != nil {
-			slog.Warn("bpf_tamper JSONL append failed", "err", err)
-		}
-	}
-}
-
-func compileEnforceAllowlist(ctx context.Context, cfg config.Config, resolver policy.LookupIPFunc, maxAttempts int) (policy.CompileResult, error) {
-	if cfg.Mode != config.ModeEnforce {
-		return policy.CompileResult{}, nil
-	}
-	if len(cfg.AllowedDomains) == 0 {
-		return policy.CompileResult{}, fmt.Errorf("enforce mode requires non-empty allowlist")
-	}
-	compiled := policy.CompileDomainAllowlist(ctx, cfg.AllowedDomains, resolver, maxAttempts)
-	if len(compiled.Domains) == 0 {
-		return policy.CompileResult{}, fmt.Errorf("enforce mode requires non-empty allowlist after normalization")
-	}
-	pol, perr := cfg.Policy()
-	if perr != nil {
-		return policy.CompileResult{}, perr
-	}
-	pol.MergeLiteralAllowedIPv4Into(&compiled.AllowedIPv4)
-	if compiled.AllowedIPv4.Len() == 0 {
-		msg := "enforce allowlist effective allowlist is empty (no IPv4 A-record resolutions; add literals to allowed-ips if needed)"
-		if len(compiled.UnresolvedDomains) > 0 {
-			msg += fmt.Sprintf(" — check DNS for: %s", strings.Join(compiled.UnresolvedDomains, ", "))
-		}
-		return policy.CompileResult{}, fmt.Errorf("%s", msg)
-	}
-	return compiled, nil
-}
-
-// loadIgnoredLPMMap programs the BPF LPM trie used to bypass denies for ignored IPv4 CIDRs.
-func loadIgnoredLPMMap(m *ebpf.Map, nets []*net.IPNet) (int, error) {
-	if len(nets) == 0 {
-		return 0, nil
-	}
-	if m == nil {
-		return 0, fmt.Errorf("ignored_ipv4_lpm map is nil with %d ignored CIDR(s)", len(nets))
-	}
-	if len(nets) > policy.MaxIgnoredIPv4Nets {
-		return 0, fmt.Errorf("ignored_ipv4_lpm: %d CIDRs exceeds max %d", len(nets), policy.MaxIgnoredIPv4Nets)
-	}
-	val := uint8(1)
-	programmed := 0
-	for i := 0; i < len(nets); i++ {
-		n := nets[i]
-		if n == nil {
-			continue
-		}
-		ones, bits := n.Mask.Size()
-		if bits != 32 || ones < 0 || ones > 32 {
-			continue
-		}
-		ip4 := n.IP.To4()
-		if ip4 == nil {
-			continue
-		}
-		network := ip4.Mask(n.Mask)
-		if network == nil {
-			continue
-		}
-		var key [8]byte
-		binary.LittleEndian.PutUint32(key[0:4], uint32(ones))
-		binary.BigEndian.PutUint32(key[4:8], binary.BigEndian.Uint32(network))
-		if err := m.Update(key, val, ebpf.UpdateAny); err != nil {
-			return 0, fmt.Errorf("ignored_ipv4_lpm update %s: %w", n.String(), err)
-		}
-		programmed++
-	}
-	if programmed == 0 {
-		return 0, fmt.Errorf(
-			"ignored_ipv4_lpm: no entries programmed from %d configured CIDR(s) (need usable IPv4 prefixes for this LPM map)",
-			len(nets),
-		)
-	}
-	return programmed, nil
-}
-
-// readUint32CounterMap reads a single-entry uint32-keyed/uint32-valued BPF counter map at key 0.
-//
-// Failure semantics (M-07): "key not found" is the legitimate zero state and is returned silently.
-// Any other Lookup error (map closed, wrong type, EBADF, program unloaded) is logged at WARN and
-// surfaced as zero so digest rendering keeps progressing — losing the distinction between "counter
-// is genuinely zero" and "map is unreadable" was the M-07 anti-pattern. The H-05 instance of this
-// pattern (enforce_cfg) is owned by Group A; this helper deliberately stays scoped to read-only
-// counter maps and never touches enforcement state.
-func readUint32CounterMap(m *ebpf.Map, helperName string) int {
-	if m == nil {
-		return 0
-	}
-	var k uint32
-	var v uint32
-	if err := m.Lookup(&k, &v); err != nil {
-		if errors.Is(err, ebpf.ErrKeyNotExist) {
-			return 0
-		}
-		slog.Warn("uint32 counter map lookup failed", "helper", helperName, "err", err)
-		return 0
-	}
-	return int(v)
-}
-
-// readUint32PerCPUArraySum sums all CPU slots for a BPF_MAP_TYPE_PERCPU_ARRAY map
-// with uint32 values at key 0. Used after migrating reserve-failure maps off a
-// contended global ARRAY slot.
-func readUint32PerCPUArraySum(m *ebpf.Map, helperName string) int {
-	if m == nil {
-		return 0
-	}
-	var k uint32
-	var vals []uint32
-	if err := m.Lookup(&k, &vals); err != nil {
-		if errors.Is(err, ebpf.ErrKeyNotExist) {
-			return 0
-		}
-		slog.Warn("percpu uint32 map lookup failed", "helper", helperName, "err", err)
-		return 0
-	}
-	n := 0
-	for _, v := range vals {
-		n += int(v)
-	}
-	return n
-}
-
-func readDenyReserveFailureCount(objs *traceenforce.TraceenforceObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.DenyReserveFailures, "readDenyReserveFailureCount")
-}
-
-func readConnect4TupleUpdateFailureCount(objs *traceconnect.TraceconnectObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.Connect4TupleUpdateFailures, "readConnect4TupleUpdateFailureCount")
-}
-
-func readUDPRingbufReserveFailureCount(objs *traceconnect.TraceconnectObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.UdpRingbufReserveFailures, "readUDPRingbufReserveFailureCount")
-}
-
-func readDNSRingbufReserveFailureCount(objs *tracedns.TracednsObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.DnsRingbufReserveFailures, "readDNSRingbufReserveFailureCount")
-}
-
-func readConnectRingbufReserveFailureCount(objs *traceconnect.TraceconnectObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.ConnectRingbufReserveFailures, "readConnectRingbufReserveFailureCount")
-}
-
-func readBPFAuditRingbufReserveFailureCount(objs *tracebpfaudit.TracebpfauditObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.BpfAuditReserveFailures, "readBPFAuditRingbufReserveFailureCount")
-}
-
-func readHTTPRingbufReserveFailureCount(objs *traceconnect.TraceconnectObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.HttpRingbufReserveFailures, "readHTTPRingbufReserveFailureCount")
-}
-
-func readTLSRingbufReserveFailureCount(objs *traceconnect.TraceconnectObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.TlsRingbufReserveFailures, "readTLSRingbufReserveFailureCount")
-}
-
-func readUDPSendmsgMultiIovecObservedCount(objs *traceconnect.TraceconnectObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32CounterMap(objs.UdpSendmsgMultiIovecObserved, "readUDPSendmsgMultiIovecObservedCount")
-}
-
-func readTLSWritevMultiIovecObservedCount(objs *traceconnect.TraceconnectObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32CounterMap(objs.TlsWritevMultiIovecObserved, "readTLSWritevMultiIovecObservedCount")
-}
-
-func readUnobservedEgressSyscallsCount(objs *traceconnect.TraceconnectObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32CounterMap(objs.UnobservedEgressSyscallsObserved, "readUnobservedEgressSyscallsCount")
-}
-
-func readIoUringSetupObservedCount(objs *traceconnect.TraceconnectObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32CounterMap(objs.IoUringSetupObserved, "readIoUringSetupObservedCount")
-}
-
-func readTCPDNSResponsesObservedCount(objs *tracedns.TracednsObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32CounterMap(objs.TcpDnsResponsesObserved, "readTCPDNSResponsesObservedCount")
-}
-
-func readTCPDNSSkippedShortReadCount(objs *tracedns.TracednsObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32CounterMap(objs.TcpDnsSkippedShortRead, "readTCPDNSSkippedShortReadCount")
-}
-
-func readExecRingbufReserveFailureCount(objs *traceexec.TraceexecObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.ExecRingbufReserveFailures, "readExecRingbufReserveFailureCount")
-}
-
-func readForkRingbufReserveFailureCount(objs *tracefork.TraceforkObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.ForkRingbufReserveFailures, "readForkRingbufReserveFailureCount")
-}
-
-func readFSRingbufReserveFailureCount(objs *tracefs.TracefsObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.FsRingbufReserveFailures, "readFSRingbufReserveFailureCount")
-}
-
-func readLSMDenyReserveFailureCount(objs *tracelsmenforce.TracelsmenforceObjects) int {
-	if objs == nil {
-		return 0
-	}
-	return readUint32PerCPUArraySum(objs.LsmDenyReserveFailures, "readLSMDenyReserveFailureCount")
-}
-
-func loadLSMEnforceMaps(objs *tracelsmenforce.TracelsmenforceObjects, compiled policy.CompileResult, pol *policy.Policy) (int, int, error) {
-	if objs == nil {
-		return 0, 0, fmt.Errorf("tracelsmenforce objects are required for enforce mode")
-	}
-	keyMode := uint32(0)
-	modeEnforce := uint32(1)
-	if err := objs.LsmEnforceCfg.Update(&keyMode, &modeEnforce, ebpf.UpdateAny); err != nil {
-		return 0, 0, fmt.Errorf("load lsm_enforce_cfg map: %w", err)
-	}
-	ignoredCount := 0
-	if pol != nil {
-		var err error
-		ignoredCount, err = loadIgnoredLPMMap(objs.LsmIgnoredIpv4Lpm, pol.IgnoredIPv4Nets())
-		if err != nil {
-			return 0, 0, err
-		}
-	}
-
-	v4keys := make(map[[4]byte]struct{}, compiled.AllowedIPv4.Len())
-	compiled.AllowedIPv4.ForEach(func(k [4]byte) { v4keys[k] = struct{}{} })
-	if pol != nil {
-		pol.MergeLiteralAllowedIPv4Keys(v4keys)
-	}
-	var literalNets []*net.IPNet
-	if pol != nil {
-		literalNets = pol.AllowedIPv4Nets()
-	}
-	totalEntries := len(v4keys) + len(literalNets)
-	if totalEntries > policy.MaxAllowedEnforceIPv4Keys {
-		return 0, 0, fmt.Errorf("lsm_allowed_ipv4: %d entries exceeds BPF max %d", totalEntries, policy.MaxAllowedEnforceIPv4Keys)
-	}
-
-	if totalEntries == 0 {
-		return 0, 0, fmt.Errorf("enforce allowlist effective allowlist is empty (no map entries)")
-	}
-
-	if err := loadAllowedLPMMap(objs.LsmAllowedIpv4, v4keys, literalNets); err != nil {
-		return 0, 0, err
-	}
-
-	if err := loadAllowedDomainsMap(objs.AllowedDomains, pol); err != nil {
-		return 0, 0, err
-	}
-
-	return totalEntries, ignoredCount, nil
-}
-
-// loadEnforceMaps programs BPF allowlist maps from compiled domain resolutions + literal policy entries.
-//
-// PR-G: allowed_ipv4 is now a BPF_MAP_TYPE_LPM_TRIE (was HASH). Single-IP
-// allowlist entries (resolved domain IPs + literal /32s from --allowed-ips)
-// are still programmed individually but with prefixlen=32. Literal CIDR
-// entries from --allowed-ips (e.g. "10.0.0.0/8") are programmed once as a
-// single LPM key and cover every address inside the range.
-func loadEnforceMaps(objs *traceenforce.TraceenforceObjects, compiled policy.CompileResult, pol *policy.Policy) (int, int, error) {
-	if objs == nil {
-		return 0, 0, fmt.Errorf("traceenforce objects are required for enforce mode")
-	}
-	keyMode := uint32(0)
-	modeEnforce := uint32(1)
-	if err := objs.EnforceCfg.Update(&keyMode, &modeEnforce, ebpf.UpdateAny); err != nil {
-		return 0, 0, fmt.Errorf("load enforce_cfg map: %w", err)
-	}
-	ignoredCount := 0
-	if pol != nil {
-		var err error
-		ignoredCount, err = loadIgnoredLPMMap(objs.IgnoredIpv4Lpm, pol.IgnoredIPv4Nets())
-		if err != nil {
-			return 0, 0, err
-		}
-	}
-
-	v4keys := make(map[[4]byte]struct{}, compiled.AllowedIPv4.Len())
-	compiled.AllowedIPv4.ForEach(func(k [4]byte) { v4keys[k] = struct{}{} })
-	if pol != nil {
-		pol.MergeLiteralAllowedIPv4Keys(v4keys)
-	}
-	var literalNets []*net.IPNet
-	if pol != nil {
-		literalNets = pol.AllowedIPv4Nets()
-	}
-	totalEntries := len(v4keys) + len(literalNets)
-	if totalEntries > policy.MaxAllowedEnforceIPv4Keys {
-		return 0, 0, fmt.Errorf("allowed_ipv4: %d entries exceeds BPF max %d", totalEntries, policy.MaxAllowedEnforceIPv4Keys)
-	}
-
-	if totalEntries == 0 {
-		return 0, 0, fmt.Errorf("enforce allowlist effective allowlist is empty (no map entries)")
-	}
-
-	if err := loadAllowedLPMMap(objs.AllowedIpv4, v4keys, literalNets); err != nil {
-		return 0, 0, err
-	}
-
-	if err := loadAllowedDomainsMap(objs.AllowedDomains, pol); err != nil {
-		return 0, 0, err
-	}
-
-	return totalEntries, ignoredCount, nil
-}
-
-// loadAllowedLPMMap programs the allowed_ipv4 LPM trie (PR-G).
-//
-// Two-phase fill keeps the kernel call sequence deterministic for tests:
-//  1. Single-IP keys (resolved domain IPs + literal /32s) → prefixlen=32.
-//  2. Literal CIDRs from --allowed-ips → prefixlen from the mask.
-//
-// Key wire format mirrors loadIgnoredLPMMap: 8-byte buffer where bytes [0:4]
-// are the prefix length in CPU/little-endian order (BPF_MAP_TYPE_LPM_TRIE
-// reads it as a u32) and bytes [4:8] are the network address in network byte
-// order. Don't reorder fields without also updating the BPF `struct ns_lpm4_key`
-// definition in bpf/trace_enforce.bpf.c — they share wire format.
-func loadAllowedLPMMap(m *ebpf.Map, ipKeys map[[4]byte]struct{}, nets []*net.IPNet) error {
-	if m == nil {
-		if len(ipKeys) > 0 || len(nets) > 0 {
-			return fmt.Errorf("allowed_ipv4 map is nil with %d entries", len(ipKeys)+len(nets))
-		}
-		return nil
-	}
-	val := uint8(1)
-	for addr := range ipKeys {
-		var key [8]byte
-		binary.LittleEndian.PutUint32(key[0:4], 32)
-		copy(key[4:8], addr[:])
-		if err := m.Update(key, val, ebpf.UpdateAny); err != nil {
-			return fmt.Errorf("load allowed_ipv4 map (/32 %d.%d.%d.%d): %w",
-				addr[0], addr[1], addr[2], addr[3], err)
-		}
-	}
-	for _, n := range nets {
-		if n == nil {
-			slog.Warn("allowed_ipv4: skipping nil CIDR entry in allowed_ipv4 LPM load")
-			continue
-		}
-		ones, bits := n.Mask.Size()
-		if bits != 32 || ones < 0 || ones > 32 {
-			slog.Warn("allowed_ipv4: skipping CIDR with non-IPv4 mask (unexpected from policy parse)", "cidr", n.String(), "bits", bits, "ones", ones)
-			continue
-		}
-		ip4 := n.IP.To4()
-		if ip4 == nil {
-			slog.Warn("allowed_ipv4: skipping non-IPv4 CIDR (unexpected from policy parse)", "cidr", n.String())
-			continue
-		}
-		network := ip4.Mask(n.Mask)
-		if network == nil {
-			slog.Warn("allowed_ipv4: skipping CIDR with nil masked network (unexpected)", "cidr", n.String())
-			continue
-		}
-		var key [8]byte
-		binary.LittleEndian.PutUint32(key[0:4], uint32(ones))
-		binary.BigEndian.PutUint32(key[4:8], binary.BigEndian.Uint32(network))
-		if err := m.Update(key, val, ebpf.UpdateAny); err != nil {
-			return fmt.Errorf("load allowed_ipv4 map (cidr %s): %w", n.String(), err)
-		}
-	}
-	return nil
-}
-
-func appendDenyFromRaw(cfg config.Config, raw []byte, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, state *enforcementState, signer *telemetry.Signer) (telemetry.DenyEvent, error) {
-	tgid, tid, commb, protocolRaw, reasonRaw, af, daddr16, dport, ok := decodeDenyEvent(raw)
-	if !ok {
-		return telemetry.DenyEvent{}, fmt.Errorf("decode deny event")
-	}
-	protocol := denyProtocolLabel(protocolRaw)
-	reason := denyReasonLabel(reasonRaw)
-	if af != linuxAFInet {
-		return telemetry.DenyEvent{}, fmt.Errorf("deny event: unsupported address family %d (IPv4 only)", af)
-	}
-	dst := net.IPv4(daddr16[0], daddr16[1], daddr16[2], daddr16[3]).String()
-	comm := string(bytes.TrimRight(commb[:], "\x00"))
-	ts := time.Now().UTC().Format(time.RFC3339Nano)
-	// Build the deny event without Seq up front; Seq is only assigned when the JSONL writer
-	// branch fires, so it stays paired with the actual JSONL line under jsonlMu (M-05) and
-	// avoids burning sequence numbers when EventsLogPath is empty (M-06). Other JSONL writers
-	// follow the same lock-then-Seq.Next() pattern (e.g. exec/tcp paths).
-	deny := telemetry.DenyEvent{
-		Type:     "deny",
-		TS:       ts,
-		PID:      tgid,
-		TGID:     tgid,
-		ThreadID: tid,
-		Comm:     comm,
-		Protocol: protocol,
-		Dst:      dst,
-		Dport:    dport,
-		Reason:   reason,
-		Mode:     cfg.PublicMode(),
-	}
-	if cfg.EventsLogPath != "" {
-		jsonlMu.Lock()
-		deny.Seq = seq.Next()
-		err := telemetry.AppendJSONL(cfg.EventsLogPath, deny, signer)
-		jsonlMu.Unlock()
-		if err != nil {
-			return telemetry.DenyEvent{}, fmt.Errorf("append deny jsonl: %w", err)
-		}
-	}
-	if state != nil {
-		state.noteDeny(denyDigestRowFromEvent(deny))
-	}
-	return deny, nil
-}
-
-// testAppendDenySample exercises appendDenyFromRaw JSONL emission and returns a sentinel error
-// for unit tests. Production readDenyRing logs and skips decode/JSONL failures so enforcement
-// keeps running; successful denies still flow through appendDenyFromRaw unchanged.
-func testAppendDenySample(cfg config.Config, raw []byte, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, state *enforcementState, signer *telemetry.Signer) error {
-	deny, err := appendDenyFromRaw(cfg, raw, seq, jsonlMu, state, signer)
-	if err != nil {
-		return err
-	}
-	return newEnforceDenyError(deny)
-}
-
-func readDenyRing(ctx context.Context, cfg config.Config, rd *ringbuf.Reader, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, state *enforcementState, signer *telemetry.Signer) error {
-	// Long-running deny consumer: drain a short burst per kernel wakeup for JSONL, then keep
-	// reading. Do not fail-fast exit on the first deny — background egress on hosted runners can
-	// emit denies immediately while the GitHub Action is still polling .coldstep-ready.json, which
-	// would kill the agent before later job steps (nmap/curl) run. Exit only on ctx cancel / closed ring.
-	backoff := newRingReadRetryBackoff()
-	drainBackoff := newRingReadRetryBackoff()
-	for {
-		rd.SetDeadline(time.Time{})
-		record, err := rd.Read()
-		if err != nil {
-			if errors.Is(err, ringbuf.ErrClosed) {
-				return nil
-			}
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			delay := backoff.sleep()
-			slog.Warn("ringbuf read (deny)", "err", err, "backoff", delay)
-			continue
-		}
-		backoff.reset()
-
-		if _, err := appendDenyFromRaw(cfg, record.RawSample, seq, jsonlMu, state, signer); err != nil {
-			slog.Warn("deny ring sample skipped", "err", err)
-			continue
-		}
-
-		drainUntil := time.Now().Add(enforceDenyDrainDuration)
-		n := 1
-		for n < enforceDenyDrainMaxEvents && time.Now().Before(drainUntil) {
-			rd.SetDeadline(time.Now().Add(enforceDenyDrainReadSlice))
-			rec2, err2 := rd.Read()
-			if err2 != nil {
-				if errors.Is(err2, ringbuf.ErrClosed) {
-					break
-				}
-				if errors.Is(err2, os.ErrDeadlineExceeded) {
-					continue
-				}
-				if ctx.Err() != nil {
-					rd.SetDeadline(time.Time{})
-					return ctx.Err()
-				}
-				rd.SetDeadline(time.Time{})
-				delay := drainBackoff.sleep()
-				slog.Warn("ringbuf read (deny drain)", "err", err2, "backoff", delay)
-				continue
-			}
-			drainBackoff.reset()
-			if _, err3 := appendDenyFromRaw(cfg, rec2.RawSample, seq, jsonlMu, state, signer); err3 != nil {
-				slog.Warn("deny ring drain sample skipped", "err", err3)
-				continue
-			}
-			n++
-		}
-		rd.SetDeadline(time.Time{})
-	}
-}
-
-// processDenyRingSample handles one deny ringbuf payload. Decode or JSONL failures are logged and
-// dropped so readDenyRing never returns a fatal error (enforcement stays attached).
-func processDenyRingSample(cfg config.Config, raw []byte, seq *telemetry.SeqGen, jsonlMu *sync.Mutex, state *enforcementState, signer *telemetry.Signer) {
-	deny, err := appendDenyFromRaw(cfg, raw, seq, jsonlMu, state, signer)
-	if err != nil {
-		slog.Warn("deny ring sample skipped", "err", err, "raw_len", len(raw))
-		return
-	}
-	slog.Debug("enforce deny", "protocol", deny.Protocol, "dst", deny.Dst, "dport", deny.Dport,
-		"reason", deny.Reason, "comm", deny.Comm)
-}
-
-func preferRunError(current error, candidate error) error {
-	if candidate == nil || errors.Is(candidate, context.Canceled) {
-		return current
-	}
-	if current == nil {
-		return candidate
-	}
-	if isEnforceDenyError(candidate) && !isEnforceDenyError(current) {
-		return candidate
-	}
-	return current
-}
-
-func bpfDetail(err error) string {
-	if err == nil {
-		return ""
-	}
-	s := err.Error()
-	const max = 180
-	if len(s) <= max {
-		return s
-	}
-	return report.TruncateUTF8ToMaxBytes(s, max) + "…"
-}
-
-func hookDegraded(bpf []telemetry.BPFStatus, hookName string) bool {
-	for _, row := range bpf {
-		if row.Name == hookName {
-			return !row.OK
-		}
-	}
-	return true
-}
-
-func capabilityEnabled(gate bool, bpf []telemetry.BPFStatus, hookName string) bool {
-	return gate && !hookDegraded(bpf, hookName)
-}
-
-// digestEnforcementLabel maps internal enforcement snapshot + config to the digest/JSONL-facing mode name.
-func digestEnforcementLabel(cfg config.Config, snap enforcementSnapshot) string {
-	if cfg.Mode != config.ModeEnforce {
-		return snap.mode
-	}
-	if strings.TrimSpace(snap.mode) != "" {
-		return snap.mode
-	}
-	return "defend"
-}
-
-func buildDigestInput(
-	cfg config.Config,
-	stats *runStats,
-	bpfSt []telemetry.BPFStatus,
-	execRows []report.ExecDigestRow,
-	tcpRows []report.TCPDigestRow,
-	udpRows []report.UDPDigestRow,
-	httpRows []report.HTTPDigestRow,
-	tlsRows []report.TLSDigestRow,
-	jsonlPath string,
-	seqLast uint64,
-	maxRows int,
-	sectionState networkSectionSnapshot,
-	enforceState enforcementSnapshot,
-	forkEdges []proctree.Edge,
-	forkEdgesTrunc bool,
-	forkSnap forkSectionSnapshot,
-	procTreeGate bool,
-	tlsSNIGate bool,
-	fsRows []report.FSDigestRow,
-	fsSnap fsSectionSnapshot,
-	fsGate bool,
-	canarySnap canarySnapshot,
-) report.DigestInput {
-	execN, tcpN, udpN, httpN, tlsN, fsN := stats.counts()
-	rawTPName := "raw_tp/sys_enter (connect, sendto, http sniff, tls)"
-	in := report.DigestInput{
-		DetectProfile:                  cfg.DetectProfile,
-		BPF:                            bpfSt,
-		ExecTotal:                      execN,
-		TCPTotal:                       tcpN,
-		UDPTotal:                       udpN,
-		HTTPTotal:                      httpN,
-		TLSTotal:                       tlsN,
-		TLSSNIGate:                     tlsSNIGate,
-		PolicyCounts:                   stats.snapshotPolicyCounts(),
-		ExecRows:                       execRows,
-		TCPRows:                        tcpRows,
-		UDPRows:                        udpRows,
-		HTTPRows:                       httpRows,
-		TLSRows:                        tlsRows,
-		JSONLPath:                      jsonlPath,
-		SeqFirst:                       1,
-		SeqLast:                        seqLast,
-		MaxRowsPerSection:              maxRows,
-		TruncatedExec:                  execN > maxRows,
-		TruncatedTCP:                   tcpN > maxRows,
-		TruncatedUDP:                   udpN > maxRows,
-		TruncatedHTTP:                  httpN > maxRows,
-		TruncatedTLS:                   tlsN > maxRows,
-		TCPDegradedHook:                hookDegraded(bpfSt, rawTPName),
-		TCPReaderErrors:                sectionState.tcpReadErrors + sectionState.tcpDecodeErrors,
-		UDPDegradedHook:                hookDegraded(bpfSt, rawTPName),
-		UDPReaderErrors:                sectionState.udpReadErrors + sectionState.udpDecodeErrors,
-		HTTPDegradedHook:               hookDegraded(bpfSt, rawTPName),
-		HTTPReaderErrors:               sectionState.httpReadErrors + sectionState.httpDecodeErrors,
-		TLSDegradedHook:                hookDegraded(bpfSt, rawTPName),
-		TLSReaderErrors:                sectionState.tlsReadErrors + sectionState.tlsDecodeErrors,
-		EnforcementMode:                digestEnforcementLabel(cfg, enforceState),
-		EnforcementAllowlistSize:       enforceState.allowlistSize,
-		EnforcementDenyCount:           enforceState.denyCount,
-		EnforcementDenyReserveFailures: enforceState.denyReserveFailures,
-		EnforcementFirstDeny:           enforceState.firstDeny,
-		Connect4TupleUpdateFailures:    stats.connect4TupleUpdateFailures(),
-		UDPRingbufReserveFailures:      stats.udpRingbufReserveFailures(),
-		DNSRingbufReserveFailures:      stats.dnsRingbufReserveFailures(),
-		ConnectRingbufReserveFailures:  stats.connectRingbufReserveFailures(),
-		HTTPRingbufReserveFailures:     stats.httpRingbufReserveFailures(),
-		TLSRingbufReserveFailures:      stats.tlsRingbufReserveFailures(),
-		ExecRingbufReserveFailures:     stats.execRingbufReserveFailures(),
-		ForkRingbufReserveFailures:     stats.forkRingbufReserveFailures(),
-		FSRingbufReserveFailures:       stats.fsRingbufReserveFailures(),
-		UDPSendmsgMultiIovecObserved:   stats.udpSendmsgMultiIovecObserved(),
-		TLSWritevMultiIovecObserved:    stats.tlsWritevMultiIovecObserved(),
-		UnobservedEgressSyscalls:       stats.unobservedEgressSyscalls(),
-		IoUringSetupObserved:           stats.ioUringSetupObserved(),
-		CanaryPipelineOK:               canarySnap.pipelineOK,
-		CanaryFailCount:                canarySnap.failCount,
-		TCPDNSResponsesObserved:        stats.tcpDNSResponsesObserved(),
-		TCPDNSSkippedShortRead:         stats.tcpDNSSkippedShortRead(),
-		BPFAuditTotal:                  stats.bpfAuditTotal(),
-		BPFMapIntegrityFailures:        stats.bpfMapIntegrityFailures(),
-		BPFAuditRingbufReserveFailures: stats.bpfAuditRingbufReserveFailures(),
-		BPFHeartbeatFailures:           stats.bpfHeartbeatFailureCount(),
-		DroppedCounts:                  stats.snapshotDroppedCounts(),
-		FSGate:                         fsGate,
-		FSTotal:                        fsN,
-		FSRows:                         fsRows,
-		TruncatedFS:                    fsN > maxRows,
-		FSDegradedHook:                 fsGate && hookDegraded(bpfSt, "raw_tp/sys_enter (fs)"),
-		FSReaderErrors:                 fsSnap.readErrors,
-	}
-	if procTreeGate {
-		in.ProcForkTotal = stats.procForkTotal()
-		in.ProcForkDegraded = hookDegraded(bpfSt, "sched_process_fork")
-		in.ProcForkReaderErrors = forkSnap.readErrors
-		in.TruncatedProcessTree = forkEdgesTrunc
-		execID := make(map[uint32]proctree.ExecIdentity, len(execRows)+8)
-		for _, r := range execRows {
-			execID[r.PID] = proctree.ExecIdentity{Comm: r.Comm, Exe: r.Exe}
-		}
-		in.ProcessTreeLines = proctree.FormatForestLines(forkEdges, execID, maxRows)
-	}
-	if seqLast == 0 {
-		in.SeqFirst = 0
-	}
-	return in
-}
 
 // Run loads BPF, streams events until ctx is cancelled, then drains workers.
 func Run(ctx context.Context, cfg config.Config) error {
